refactor: Complete migration to modular build system and production Tailwind
BREAKING CHANGES:
- Removed build-static.js in favor of modular build.js
- Removed monolithic template/index.html
- Consolidated enhanced JSON files into main ones
- Removed venues.json (won't be maintained)

Features:
- Set up production Tailwind CSS using CLI (25KB minified)
- Removed Tailwind CDN dependency
- Migrated entirely to Handlebars modular template system
- Simplified data structure with single JSON files

Cleanup:
- Renamed build-modular.js to build.js
- Updated package.json scripts
- Removed all test/backup HTML files
- Updated CLAUDE.md documentation
- Created comprehensive TODO.md with prioritized tasks
