--- public/data/venues.json
@@ -1,144 +1,0 @@
-{
-  "venues": {
-    "club-fox": {
-      "id": "club-fox",
-      "name": "Club Fox",
-      "website": "https://clubfoxrwc.com/",
-      "phone": "+1-650-369-7770",
-      "address": {
-        "streetAddress": "2209 Broadway",
-        "city": "Redwood City",
-        "state": "CA",
-        "postalCode": "94063",
-        "country": "US"
-      },
-      "geo": {
-        "latitude": 37.4866,
-        "longitude": -122.2284
-      },
-      "capacity": 300,
-      "type": "Live Music Venue",
-      "amenities": [
-        "Full Bar",
-        "Food Menu",
-        "Dance Floor",
-        "Professional Sound System",
-        "Stage Lighting"
-      ],
-      "accessibility": {
-        "wheelchairAccessible": true,
-        "assistiveListening": true,
-        "parkingAvailable": true,
-        "publicTransit": "Caltrain Redwood City Station (0.3 miles)"
-      },
-      "parkingInfo": "Street parking and nearby public lots",
-      "ageRestriction": "21+",
-      "description": "Intimate live music venue in downtown Redwood City featuring local and touring acts",
-      "images": {
-        "exterior": "club-fox-exterior.jpg",
-        "interior": "club-fox-interior.jpg",
-        "stage": "club-fox-stage.jpg"
-      }
-    },
-    "stafford-park": {
-      "id": "stafford-park",
-      "name": "Stafford Park",
-      "website": "https://www.redwoodcity.org/departments/parks-recreation-and-community-services/parks",
-      "address": {
-        "streetAddress": "Corner of King St. & Hopkins Ave",
-        "city": "Redwood City",
-        "state": "CA",
-        "postalCode": "94061",
-        "country": "US"
-      },
-      "geo": {
-        "latitude": 37.4848,
-        "longitude": -122.2281
-      },
-      "capacity": 2000,
-      "type": "Outdoor Park",
-      "amenities": [
-        "Large Lawn Area",
-        "Playground",
-        "Picnic Tables",
-        "Restrooms",
-        "Food Trucks (during events)"
-      ],
-      "accessibility": {
-        "wheelchairAccessible": true,
-        "parkingAvailable": true,
-        "publicTransit": "Caltrain Redwood City Station (0.5 miles)"
-      },
-      "parkingInfo": "Free street parking, ADA parking available",
-      "ageRestriction": "All Ages",
-      "description": "Community park hosting the annual Music in the Park summer concert series"
-    },
-    "blondies-bar": {
-      "id": "blondies-bar",
-      "name": "Blondie's Bar",
-      "website": "https://blondiesbarsf.com/",
-      "phone": "+1-415-864-2419",
-      "address": {
-        "streetAddress": "540 Valencia St",
-        "city": "San Francisco",
-        "state": "CA",
-        "postalCode": "94110",
-        "country": "US"
-      },
-      "geo": {
-        "latitude": 37.7644,
-        "longitude": -122.4217
-      },
-      "capacity": 200,
-      "type": "Bar/Club",
-      "amenities": [
-        "Full Bar",
-        "Outdoor Patio",
-        "DJ Booth",
-        "Dance Floor"
-      ],
-      "accessibility": {
-        "wheelchairAccessible": true,
-        "parkingAvailable": false,
-        "publicTransit": "16th St Mission BART (0.2 miles)"
-      },
-      "parkingInfo": "Street parking (metered), nearby parking garages",
-      "ageRestriction": "21+",
-      "description": "Popular Mission District bar known for live music and Valencia Street events"
-    },
-    "the-fillmore": {
-      "id": "the-fillmore",
-      "name": "The Fillmore",
-      "website": "https://www.thefillmore.com/",
-      "phone": "+1-415-346-6000",
-      "address": {
-        "streetAddress": "1805 Geary Blvd",
-        "city": "San Francisco",
-        "state": "CA",
-        "postalCode": "94115",
-        "country": "US"
-      },
-      "geo": {
-        "latitude": 37.7843,
-        "longitude": -122.4330
-      },
-      "capacity": 1150,
-      "type": "Historic Music Venue",
-      "amenities": [
-        "Full Bar",
-        "Balcony Seating",
-        "Historic Poster Gallery",
-        "Professional Sound & Lighting"
-      ],
-      "accessibility": {
-        "wheelchairAccessible": true,
-        "assistiveListening": true,
-        "parkingAvailable": false,
-        "publicTransit": "Multiple MUNI lines"
-      },
-      "parkingInfo": "Limited street parking, public garage on Fillmore St",
-      "ageRestriction": "All Ages (bar 21+)",
-      "description": "Legendary San Francisco music venue with rich history in rock, soul, and funk"
-    }
-  }
-}
