--- public/index-modular.php
@@ -1,433 +1,0 @@
-<!doctype html> <html lang="en"> <head> <script>!function(e,t,a,n){e[n]=e[n]||[],e[n].push({"gtm.start":(new Date).getTime(),event:"gtm.js"});var g=t.getElementsByTagName(a)[0],m=t.createElement(a);m.async=!0,m.src="https://www.googletagmanager.com/gtm.js?id=GTM-NJQVBL6T",g.parentNode.insertBefore(m,g)}(window,document,"script","dataLayer")</script> <meta charset="UTF-8"> <meta content="width=device-width,initial-scale=1,viewport-fit=cover" name="viewport"> <title>Sinister Dexter - The Right to Be Funky | San Francisco Soul & Funk Band</title> <meta content="index, follow" name="robots"> <meta content="Sinister Dexter is an 11-piece soul machine bringing the unique sound of great funk horn bands to the San Francisco Bay Area. Featuring vocalist Rebecca Lipon." name="description"> <meta content="Sinister Dexter, funk band, soul music, San Francisco, Bay Area, Rebecca Lipon, live music, horn section" name="keywords"> <link href="https://sinisterdexter.net" rel="canonical"> <meta content="Sinister Dexter - The Right to Be Funky" property="og:title"> <meta content="11-piece soul machine bringing high-energy funk to the Bay Area" property="og:description"> <meta content="website" property="og:type"> <meta content="https://sinisterdexter.net" property="og:url"> <meta content="https://sinisterdexter.net/og-image.jpg" property="og:image"> <meta content="1200" property="og:image:width"> <meta content="630" property="og:image:height"> <meta content="summary_large_image" name="twitter:card"> <meta content="@sindexband" name="twitter:site"> <meta content="Sinister Dexter - The Right to Be Funky" name="twitter:title"> <meta content="11-piece soul machine bringing high-energy funk to the Bay Area" name="twitter:description"> <meta content="https://sinisterdexter.net/twitter-card.jpg" name="twitter:image"> <link href="favicon.ico" rel="icon" type="image/x-icon"> <link href="favicon-16x16.png" rel="icon" sizes="16x16" type="image/png"> <link href="favicon-32x32.png" rel="icon" sizes="32x32" type="image/png"> <link href="favicon-48x48.png" rel="icon" sizes="48x48" type="image/png"> <link href="favicon-64x64.png" rel="icon" sizes="64x64" type="image/png"> <link href="favicon-128x128.png" rel="icon" sizes="128x128" type="image/png"> <link href="apple-touch-icon.png" rel="apple-touch-icon" sizes="180x180"> <link href="favicon-192x192.png" rel="icon" sizes="192x192" type="image/png"> <link href="favicon-512x512.png" rel="icon" sizes="512x512" type="image/png"> <link href="manifest.json" rel="manifest"> <meta content="#9333ea" name="theme-color"> <link href="https://fonts.googleapis.com" rel="dns-prefetch"> <link href="https://fonts.gstatic.com" rel="dns-prefetch"> <link href="https://cdn.tailwindcss.com" rel="dns-prefetch"> <link href="https://cdnjs.cloudflare.com" rel="dns-prefetch"> <link href="https://img.youtube.com" rel="dns-prefetch"> <link href="https://www.youtube.com" rel="dns-prefetch"> <link href="https://cdn.sinister-dexter.com" rel="dns-prefetch"> <link href="https://fonts.googleapis.com/css2?family=Bebas+Neue&family=Inter:wght@400;600;700&display=swap" rel="preload" as="style"> <link href="https://fonts.googleapis.com" rel="preconnect"> <link href="https://fonts.gstatic.com" rel="preconnect" crossorigin> <style>.bebas{font-family:'Bebas Neue','Arial Narrow','Helvetica Condensed',Arial,sans-serif;font-weight:700;letter-spacing:.03em;font-stretch:condensed;text-transform:uppercase}.sr-only{position:absolute;width:1px;height:1px;padding:0;margin:-1px;overflow:hidden;clip:rect(0,0,0,0);white-space:nowrap;border-width:0}.focus\:not-sr-only:focus{position:static;width:auto;height:auto;padding:inherit;margin:inherit;overflow:visible;clip:auto;white-space:normal}.focus-visible\:ring-2:focus-visible{--tw-ring-offset-shadow:var(--tw-ring-inset) 0 0 0 var(--tw-ring-offset-width) var(--tw-ring-offset-color);--tw-ring-shadow:var(--tw-ring-inset) 0 0 0 calc(2px + var(--tw-ring-offset-width)) var(--tw-ring-color);box-shadow:var(--tw-ring-offset-shadow),var(--tw-ring-shadow),var(--tw-shadow,0 0 #0000)}.focus-visible\:ring-purple-400:focus-visible{--tw-ring-color:rgb(192 132 252)}.focus-visible\:outline-none:focus-visible{outline:2px solid transparent;outline-offset:2px}.gradient-text{background:linear-gradient(135deg,#f59e0b 0,#9333ea 100%);-webkit-background-clip:text;-webkit-text-fill-color:transparent;background-clip:text}.hero-overlay{background:radial-gradient(ellipse at center,transparent 0,rgba(0,0,0,.7) 100%)}.gallery-item{transition:transform .3s ease}.gallery-item:hover{transform:scale(1.05)}html{scroll-behavior:smooth}::-webkit-scrollbar{width:8px;height:8px}::-webkit-scrollbar-track{background:rgba(0,0,0,.1)}::-webkit-scrollbar-thumb{background:#9333ea;border-radius:4px}::-webkit-scrollbar-thumb:hover{background:#7c3aed}.track-item{cursor:pointer}.track-item:hover{background-color:rgba(147,51,234,.3)}.track-item.active{background-color:rgba(147,51,234,.5);border-left:3px solid #9333ea}.nav-link{position:relative;transition:color .3s ease}.nav-link::after{content:'';position:absolute;bottom:-2px;left:0;width:0;height:2px;background:#9333ea;transition:width .3s ease}.nav-link:hover::after{width:100%}</style> <script src="https://cdn.tailwindcss.com"></script> <link href="https://fonts.googleapis.com/css2?family=Bebas+Neue&family=Inter:wght@400;600;700&display=swap" rel="stylesheet"> <link href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/5.15.4/css/all.min.css" rel="stylesheet"> <script type="application/ld+json"> {
-      "@context": "https://schema.org",
-      "@type": "MusicGroup",
-      "@id": "https://sinisterdexter.net#band",
-      "name": "Sinister Dexter",
-      "url": "https://sinisterdexter.net",
-      "logo": "https://sinisterdexter.net/og-image.jpg",
-      "image": "https://sinisterdexter.net/images/large/sindex-full.jpg",
-      "description": "Sinister Dexter is an 11-piece soul machine bringing the unique sound of great funk horn bands to the San Francisco Bay Area. Featuring vocalist Rebecca Lipon.",
-      "foundingDate": "1998",
-      "foundingLocation": {
-        "@type": "Place",
-        "address": {
-          "@type": "PostalAddress",
-          "addressLocality": "San Francisco",
-          "addressRegion": "CA",
-          "addressCountry": "US"
-        }
-      },
-      "genre": ["Funk","Soul","R&amp;B","Blues"],
-      "member": [
-        {
-          "@type": "OrganizationRole",
-          "member": {
-            "@type": "Person",
-            "@id": "https://sinisterdexter.net#rebecca-lipon-weekly",
-            "name": "Rebecca Lipon Weekly",
-            "description": "Rebecca&#x27;s love of singing started with Sesame Street, much to her siblings&#x27; chagrin, and it never stopped. She was a founding member of the Off Beats, a Jazz quartet, in her late teens, studied opera and musical theater styles for over 15 years, and performed professionally with theater groups in Northern California, Boston and New York. She has been Sinister Dexter&#x27;s lead singer since 2005, and loves the mix of soul, funk and blues music.",
-            "image": "https://sinisterdexter.net/images/members/rebecca.jpg",
-            "jobTitle": "Lead Vocals",
-            
-            "sameAs": [
-            ]
-          },
-          "roleName": "Lead Vocals",
-          "startDate": "2005",
-          "memberOf": {
-            "@id": "https://sinisterdexter.net#band"
-          }
-        },
-        {
-          "@type": "OrganizationRole",
-          "member": {
-            "@type": "Person",
-            "@id": "https://sinisterdexter.net#daniel-walling",
-            "name": "Daniel Walling",
-            "description": "Daniel started playing piano like everyone else, because Mom made him take lessons. But then he got hooked, switching from classical to jazz in high school, playing in various jazz combos in college, and in 2004 switching again to funky tasty groovaliciousness with sindex. Critics have described his organ solos as &quot;really f*#$ing loud&quot;.",
-            "image": "https://sinisterdexter.net/images/members/walling.jpg",
-            "jobTitle": "Keyboard",
-            
-            "sameAs": [
-            ]
-          },
-          "roleName": "Keyboard",
-          "startDate": "2004",
-          "memberOf": {
-            "@id": "https://sinisterdexter.net#band"
-          }
-        },
-        {
-          "@type": "OrganizationRole",
-          "member": {
-            "@type": "Person",
-            "@id": "https://sinisterdexter.net#david-mitby",
-            "name": "David Mitby",
-            "description": "Keeping the groove tight",
-            "image": "https://sinisterdexter.net/images/members/mitby.jpg",
-            "jobTitle": "Bass",
-            
-            "sameAs": [
-            ]
-          },
-          "roleName": "Bass",
-          
-          "memberOf": {
-            "@id": "https://sinisterdexter.net#band"
-          }
-        },
-        {
-          "@type": "OrganizationRole",
-          "member": {
-            "@type": "Person",
-            "@id": "https://sinisterdexter.net#matt-spitz",
-            "name": "Matt Spitz",
-            "description": "Matt plays guitar, has performed in 38 states, and recently won a Grammy. As of this writing, at least one of those statements is correct. This is his second tour of duty with SinDex after a decade hiatus in New York City.",
-            "image": "https://sinisterdexter.net/images/members/spitz.jpg",
-            "jobTitle": "Guitar",
-            
-            "sameAs": [
-            ]
-          },
-          "roleName": "Guitar",
-          
-          "memberOf": {
-            "@id": "https://sinisterdexter.net#band"
-          }
-        },
-        {
-          "@type": "OrganizationRole",
-          "member": {
-            "@type": "Person",
-            "@id": "https://sinisterdexter.net#jon-wagner",
-            "name": "Jon Wagner",
-            "description": "",
-            "image": "https://sinisterdexter.net/images/members/wagner.jpg",
-            "jobTitle": "Drums",
-            
-            "sameAs": [
-            ]
-          },
-          "roleName": "Drums",
-          "startDate": "1998",
-          "memberOf": {
-            "@id": "https://sinisterdexter.net#band"
-          }
-        },
-        {
-          "@type": "OrganizationRole",
-          "member": {
-            "@type": "Person",
-            "@id": "https://sinisterdexter.net#kevin-wooley",
-            "name": "Kevin Wooley",
-            "description": "Kevin is a founding member of Sinister Dexter and the defacto manager. He&#x27;s been playing tenor sax since he was 12 years old in jazz groups, symphonic orchestras, and marching bands. He also plays clarinet and soprano sax, but Kenny G ruined that instrument for everyone. Because of his love for music from the sixties and seventies (and disgust with music from the eighties and nineties), he often thinks he was born a couple of decades too late. By day, Kevin is an Oscar-winning engineer for Industrial Light &amp; Magic.",
-            "image": "https://sinisterdexter.net/images/members/wooley.jpg",
-            "jobTitle": "Tenor Sax",
-            
-            "sameAs": [
-            ]
-          },
-          "roleName": "Tenor Sax",
-          "startDate": "1998",
-          "memberOf": {
-            "@id": "https://sinisterdexter.net#band"
-          }
-        },
-        {
-          "@type": "OrganizationRole",
-          "member": {
-            "@type": "Person",
-            "@id": "https://sinisterdexter.net#don-hoffman",
-            "name": "Don Hoffman",
-            "description": "Don&#x27;s been playing for over 20 years, and his background covers everything from jazz to musical pit orchestras, and symphony to marching bands. He&#x27;s one of the founding members of Sinister Dexter, and enjoys occasionally playing with other Bay Area bands including Foreverland, which he also helped start.",
-            "image": "https://sinisterdexter.net/images/members/hoffman.jpg",
-            "jobTitle": "Lead Trombone",
-            
-            "sameAs": [
-            ]
-          },
-          "roleName": "Lead Trombone",
-          "startDate": "1998",
-          "memberOf": {
-            "@id": "https://sinisterdexter.net#band"
-          }
-        },
-        {
-          "@type": "OrganizationRole",
-          "member": {
-            "@type": "Person",
-            "@id": "https://sinisterdexter.net#josh-schiller",
-            "name": "Josh Schiller",
-            "description": "Josh has been playing bone for a very long time. You could even say it pays him more than anything else he does, especially since he&#x27;s currently enjoying a life of leisure.",
-            "image": "https://sinisterdexter.net/images/members/schiller.jpg",
-            "jobTitle": "Bass Trombone",
-            
-            "sameAs": [
-            ]
-          },
-          "roleName": "Bass Trombone",
-          "startDate": "1998",
-          "memberOf": {
-            "@id": "https://sinisterdexter.net#band"
-          }
-        },
-        {
-          "@type": "OrganizationRole",
-          "member": {
-            "@type": "Person",
-            "@id": "https://sinisterdexter.net#patrick-neschleba",
-            "name": "Patrick Neschleba",
-            "description": "",
-            "image": "https://sinisterdexter.net/images/members/patrick.jpg",
-            "jobTitle": "Alto &amp; Bari Sax",
-            
-            "sameAs": [
-            ]
-          },
-          "roleName": "Alto &amp; Bari Sax",
-          "startDate": "1998",
-          "memberOf": {
-            "@id": "https://sinisterdexter.net#band"
-          }
-        },
-        {
-          "@type": "OrganizationRole",
-          "member": {
-            "@type": "Person",
-            "@id": "https://sinisterdexter.net#adam-elman",
-            "name": "Adam Elman",
-            "description": "Adam is a founding member of Sinister Dexter and is glad to be playing with them again after a brief 14-year hiatus. Starting as a trombone player in bands and orchestras in 9th grade, Adam picked up trumpet after college and decided to stick with it when SinDex began, both because the band already had two great trombone players and because it&#x27;s way easier to carry a trumpet around. In the rare moments when he&#x27;s not busy helping raise a kid or designing voice experiences for a large tech company, he likes to arrange and occasionally compose original music.",
-            "image": "https://sinisterdexter.net/images/members/elman.jpg",
-            "jobTitle": "Trumpet",
-            
-            "sameAs": [
-            ]
-          },
-          "roleName": "Trumpet",
-          "startDate": "1998",
-          "memberOf": {
-            "@id": "https://sinisterdexter.net#band"
-          }
-        },
-        {
-          "@type": "OrganizationRole",
-          "member": {
-            "@type": "Person",
-            "@id": "https://sinisterdexter.net#lori-karns",
-            "name": "Lori Karns",
-            "description": "",
-            "image": "https://sinisterdexter.net/images/members/lori.jpg",
-            "jobTitle": "Trumpet",
-            
-            "sameAs": [
-            ]
-          },
-          "roleName": "Trumpet",
-          
-          "memberOf": {
-            "@id": "https://sinisterdexter.net#band"
-          }
-        }
-      ],
-      "sameAs": [
-        "https://www.instagram.com/sindex_band",
-        "https://x.com/sindexband",
-        "https://www.facebook.com/SinDex.Band",
-        "https://www.youtube.com/@SinDexBand",
-        "https://open.spotify.com/artist/06ZpJFLW4Kz4EAlUorzrFu",
-        "https://music.apple.com/us/artist/sinister-dexter/30663361"
-      ],
-      "contactPoint": {
-        "@type": "ContactPoint",
-        "contactType": "Booking",
-        "email": "booking@sinisterdexter.net"
-      }
-    } </script> <script type="application/ld+json"> {
-      "@context": "https://schema.org",
-      "@type": "ItemList",
-      "name": "Upcoming Sinister Dexter Shows",
-      "numberOfItems": 3,
-      "itemListElement": [
-        {
-          "@type": "ListItem",
-          "position": 1,
-          "item": {
-            "@type": "MusicEvent",
-            "@id": "https://sinisterdexter.net#event-2025-08-20-redwood-city-music-in-the-park---stafford-park",
-            "name": "Sinister Dexter at Redwood City Music in the Park - Stafford Park",
-            "description": "Free summer concert series at the corner of King St. &amp; Hopkins Ave. Kid friendly!",
-            "startDate": "2025-08-20T5:00 PM",
-            "endDate": "2025-08-20T7:00 PM-08:00",
-            "location": {
-              "@type": "Place",
-              "name": "Redwood City Music in the Park - Stafford Park",
-              "address": {
-                "@type": "PostalAddress",
-                
-                "addressLocality": "Redwood City, CA",
-                "addressRegion": "CA",
-                
-                "addressCountry": ""
-              }
-            },
-            "performer": {
-              "@type": "MusicGroup",
-              "@id": "https://sinisterdexter.net#band"
-            },
-            "eventStatus": "https://schema.org/EventScheduled",
-            "eventAttendanceMode": "https://schema.org/OfflineEventAttendanceMode",
-            
-            "url": "https://www.redwoodcity.org/residents/redwood-city-events/music/music-in-the-park"
-          }
-        },
-        {
-          "@type": "ListItem",
-          "position": 2,
-          "item": {
-            "@type": "MusicEvent",
-            "@id": "https://sinisterdexter.net#event-2025-09-11-valencia-live!-at-blondie&#x27;s-bar",
-            "name": "Sinister Dexter at Valencia LIVE! at Blondie&#x27;s Bar",
-            "description": "Big night market on Valencia St with food, drink, entertainment, and FUNK!",
-            "startDate": "2025-09-11T5:00 PM",
-            "endDate": "2025-09-11T8:00 PM-08:00",
-            "location": {
-              "@type": "Place",
-              "name": "Valencia LIVE! at Blondie&#x27;s Bar",
-              "address": {
-                "@type": "PostalAddress",
-                
-                "addressLocality": "San Francisco, CA",
-                "addressRegion": "CA",
-                
-                "addressCountry": ""
-              }
-            },
-            "performer": {
-              "@type": "MusicGroup",
-              "@id": "https://sinisterdexter.net#band"
-            },
-            "eventStatus": "https://schema.org/EventScheduled",
-            "eventAttendanceMode": "https://schema.org/OfflineEventAttendanceMode",
-            
-            "url": "https://blondiesbarsf.com/"
-          }
-        },
-        {
-          "@type": "ListItem",
-          "position": 3,
-          "item": {
-            "@type": "MusicEvent",
-            "@id": "https://sinisterdexter.net#event-2025-11-14-club-fox",
-            "name": "Sinister Dexter at Club Fox",
-            "description": "Friday night funk at the legendary Club Fox venue",
-            "startDate": "2025-11-14T6:00 PM",
-            "endDate": "2025-11-14T10:00 PM-08:00",
-            "location": {
-              "@type": "Place",
-              "name": "Club Fox",
-              "address": {
-                "@type": "PostalAddress",
-                
-                "addressLocality": "Redwood City, CA",
-                "addressRegion": "CA",
-                
-                "addressCountry": ""
-              }
-            },
-            "performer": {
-              "@type": "MusicGroup",
-              "@id": "https://sinisterdexter.net#band"
-            },
-            "offers": {
-              "@type": "Offer",
-              "url": "https://www.eventbrite.com/e/sinister-dexter-funk-soul-horns-tickets-1568339939979?aff&#x3D;ebdssbdestsearch",
-              "price": "0",
-              "priceCurrency": "USD",
-              "availability": "https://schema.org/InStock"
-            },
-            "eventStatus": "https://schema.org/EventScheduled",
-            "eventAttendanceMode": "https://schema.org/OfflineEventAttendanceMode",
-            
-            "url": "https://clubfoxrwc.com/"
-          }
-        }
-      ]
-    } </script> <script type="application/ld+json"> {
-      "@context": "https://schema.org",
-      "@type": "ItemList",
-      "name": "Sinister Dexter Performance Videos",
-      "numberOfItems": 3,
-      "itemListElement": [
-        {
-          "@type": "ListItem",
-          "position": 1,
-          "item": {
-            "@type": "VideoObject",
-            "name": "Son of a Preacher Man",
-            "description": "Redwood City Music In The Park, 2012",
-            "thumbnailUrl": "https://img.youtube.com/vi/D9gIdE7IJsY/maxresdefault.jpg",
-            "uploadDate": "2012-08-15",
-            "contentUrl": "https://www.youtube.com/watch?v&#x3D;D9gIdE7IJsY",
-            "embedUrl": "https://www.youtube.com/embed/D9gIdE7IJsY",
-            "duration": "PT3M30S"
-          }
-        },
-        {
-          "@type": "ListItem",
-          "position": 2,
-          "item": {
-            "@type": "VideoObject",
-            "name": "I Can&#x27;t Stand The Rain",
-            "description": "Redwood City Music In The Park, 2012",
-            "thumbnailUrl": "https://img.youtube.com/vi/vWRJUqNdnX0/maxresdefault.jpg",
-            "uploadDate": "2012-08-15",
-            "contentUrl": "https://www.youtube.com/watch?v&#x3D;vWRJUqNdnX0",
-            "embedUrl": "https://www.youtube.com/embed/vWRJUqNdnX0",
-            "duration": "PT4M15S"
-          }
-        },
-        {
-          "@type": "ListItem",
-          "position": 3,
-          "item": {
-            "@type": "VideoObject",
-            "name": "Broadband Connection",
-            "description": "Redwood City Music In The Park, 2012",
-            "thumbnailUrl": "https://img.youtube.com/vi/buwFWtEuOyQ/maxresdefault.jpg",
-            "uploadDate": "2012-08-15",
-            "contentUrl": "https://www.youtube.com/watch?v&#x3D;buwFWtEuOyQ",
-            "embedUrl": "https://www.youtube.com/embed/buwFWtEuOyQ",
-            "duration": "PT3M01S"
-          }
-        }
-      ]
-    } </script> <script type="application/ld+json"> {
-      "@context": "https://schema.org",
-      "@type": "BreadcrumbList",
-      "itemListElement": [
-        {
-          "@type": "ListItem",
-          "position": 1,
-          "name": "Home",
-          "item": "https://sinisterdexter.net"
-        },
-        {
-          "@type": "ListItem",
-          "position": 2,
-          "name": "Shows",
-          "item": "https://sinisterdexter.net#shows"
-        },
-        {
-          "@type": "ListItem",
-          "position": 3,
-          "name": "Music",
-          "item": "https://sinisterdexter.net#music-library"
-        },
-        {
-          "@type": "ListItem",
-          "position": 4,
-          "name": "About",
-          "item": "https://sinisterdexter.net#about"
-        }
-      ]
-    } </script></head> <body class="bg-black font-sans text-white"> <noscript><iframe height="0" src="https://www.googletagmanager.com/ns.html?id=GTM-NJQVBL6T" style="display:none;visibility:hidden" title="Google Tag Manager" width="0"></iframe></noscript> <a class="rounded-lg bg-purple-600 focus:absolute focus:left-4 focus:not-sr-only focus:top-4 px-4 py-2 sr-only text-white z-[60]" href="#main-content"> Skip to main content </a> <nav aria-label="Main navigation" class="px-6 backdrop-blur-md bg-black/90 fixed py-4 top-0 w-full z-50" role="navigation"> <div class="flex items-center justify-between max-w-7xl mx-auto"> <a class="bebas gradient-text text-3xl" href="#hero">Sinister Dexter</a> <div class="hidden md:flex space-x-8"> <a class="rounded focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-purple-400 hover:text-purple-400 nav-link" href="#shows">Shows</a> <a class="rounded focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-purple-400 hover:text-purple-400 nav-link" href="#music-library">Music</a> <a class="rounded focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-purple-400 hover:text-purple-400 nav-link" href="#videos">Videos</a> <a class="rounded focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-purple-400 hover:text-purple-400 nav-link" href="#gallery">Gallery</a> <a class="rounded focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-purple-400 hover:text-purple-400 nav-link" href="#about">About</a> <a class="rounded focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-purple-400 hover:text-purple-400 nav-link" href="#band-members">Band</a> <a class="rounded focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-purple-400 hover:text-purple-400 nav-link" href="#notable-venues">Venues</a> <a class="rounded focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-purple-400 hover:text-purple-400 nav-link" href="#connect">Connect</a> </div> <button class="md:hidden" aria-label="Toggle navigation menu" id="mobile-menu-btn" aria-controls="mobile-menu" aria-expanded="false"> <svg class="h-6 w-6" fill="none" viewBox="0 0 24 24" aria-hidden="true" xmlns="http://www.w3.org/2000/svg" stroke="currentColor" stroke-width="1.5"> <path d="M3.75 6.75h16.5M3.75 12h16.5m-16.5 5.25h16.5" stroke-linecap="round" stroke-linejoin="round"/> </svg> </button> </div> <div class="hidden md:hidden mt-4 pb-4" id="mobile-menu"> <a class="rounded focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-purple-400 hover:text-purple-400 block py-2" href="#shows">Shows</a> <a class="rounded focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-purple-400 hover:text-purple-400 block py-2" href="#music-library">Music</a> <a class="rounded focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-purple-400 hover:text-purple-400 block py-2" href="#videos">Videos</a> <a class="rounded focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-purple-400 hover:text-purple-400 block py-2" href="#gallery">Gallery</a> <a class="rounded focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-purple-400 hover:text-purple-400 block py-2" href="#about">About</a> <a class="rounded focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-purple-400 hover:text-purple-400 block py-2" href="#band-members">Band</a> <a class="rounded focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-purple-400 hover:text-purple-400 block py-2" href="#notable-venues">Venues</a> <a class="rounded focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-purple-400 hover:text-purple-400 block py-2" href="#connect">Connect</a> </div> </nav> <script>document.getElementById("mobile-menu-btn").addEventListener("click",function(){const e=document.getElementById("mobile-menu");!e.classList.contains("hidden")?(e.classList.add("hidden"),this.setAttribute("aria-expanded","false")):(e.classList.remove("hidden"),this.setAttribute("aria-expanded","true"))}),document.querySelectorAll("#mobile-menu a").forEach(e=>{e.addEventListener("click",()=>{document.getElementById("mobile-menu").classList.add("hidden"),document.getElementById("mobile-menu-btn").setAttribute("aria-expanded","false")})})</script> <main id="main-content" itemscope itemtype="https://schema.org/MusicGroup" role="main"> <meta content="https://sinisterdexter.net" itemprop="url"> <section class="flex items-center justify-center h-screen overflow-hidden relative" id="hero"> <picture class="inset-0 absolute"> <source srcset="images/large/sindex-full.webp" type="image/webp"> <img alt="Sinister Dexter full band" class="h-full object-cover w-full object-center" src="images/large/sindex-full.jpg" itemprop="image"> </picture> <div class="inset-0 absolute bg-gradient-to-br from-purple-900/60 to-amber-900/60"></div> <div class="inset-0 absolute hero-overlay"></div> <div class="text-center px-6 relative z-10"> <h1 class="bebas mb-4 lg:text-9xl md:text-8xl text-6xl" itemprop="name">Sinister Dexter</h1> <p class="font-bold mb-2 text-xl md:text-3xl text-amber-400" itemprop="slogan">The Right to Be Funky</p> <p class="mx-auto max-w-2xl mb-8 md:text-xl text-lg" itemprop="description"> 11-piece soul machine bringing the unique sound of great funk horn bands to the San Francisco Bay Area </p> <div class="flex justify-center flex-col gap-4 sm:flex-row"> <a class="transition-all rounded-full hover:scale-105 py-3 font-semibold bg-purple-600 hover:bg-purple-700 px-8" href="#shows"> SEE UPCOMING SHOWS </a> <a class="transition-all rounded-full border-2 border-amber-500 font-semibold hover:bg-amber-500 hover:scale-105 hover:text-black px-8 py-3" href="#connect"> JOIN MAILING LIST </a> </div> </div> <div class="animate-bounce -translate-x-1/2 absolute bottom-8 left-1/2 transform"> <svg class="h-6 w-6 animate-bounce" fill="none" viewBox="0 0 24 24" aria-hidden="true" xmlns="http://www.w3.org/2000/svg" stroke="currentColor" stroke-width="1.5"> <path d="m19.5 8.25-7.5 7.5-7.5-7.5" stroke-linecap="round" stroke-linejoin="round"/> </svg> </div> </section> <section class="px-6 py-20 bg-gradient-to-b from-black to-purple-950/30" id="shows" aria-labelledby="shows-heading" itemscope itemtype="https://schema.org/EventSeries"> <div class="mx-auto max-w-7xl"> <h2 class="text-center bebas gradient-text md:text-6xl text-5xl mb-12" id="shows-heading">Upcoming Shows</h2> <div class="grid gap-6 lg:grid-cols-3 md:grid-cols-2" id="shows-container"> <div class="rounded-lg transition-all bg-gray-900/50 border border-purple-500/30 hover:border-purple-400 p-6 ring-2 ring-purple-500" itemscope itemtype="https://schema.org/MusicEvent" itemprop="event"> <meta content="https://sinisterdexter.net#event-2025-08-20-redwood-city-music-in-the-park---stafford-park" itemprop="@id"> <div class="text-xs mb-2 font-bold text-purple-400">FEATURED</div> <div class="font-bold mb-2 text-amber-500"> <time datetime="2025-08-20T5:00 PM-08:00" itemprop="startDate"> WED, AUG 19 </time> </div> <div itemscope itemtype="https://schema.org/Place" itemprop="location"> <h3 class="font-bold mb-2 text-xl" itemprop="name">Redwood City Music in the Park - Stafford Park</h3> <div itemscope itemtype="https://schema.org/PostalAddress" itemprop="address"> <p class="text-gray-400 mb-2"> <span itemprop="addressLocality">Redwood City, CA</span>, <span itemprop="addressRegion">CA</span> </p><meta content="" itemprop="addressCountry"> <p></p> </div> </div> <p class="text-gray-300 mb-4"> <time datetime="2025-08-20T5:00 PM">5:00 PM AM</time> - <time datetime="2025-08-20T7:00 PM" itemprop="endDate">7:00 PM AM</time> </p> <p class="text-sm text-gray-400 mb-4" itemprop="description">Free summer concert series at the corner of King St. & Hopkins Ave. Kid friendly!</p> <div class="hidden" itemscope itemtype="https://schema.org/MusicGroup" itemprop="performer"> <meta content="Sinister Dexter" itemprop="name"> <link href="https://sinisterdexter.net" itemprop="url"> </div> <div itemscope itemtype="https://schema.org/Offer" itemprop="offers"> <meta content="0" itemprop="price"> <meta content="USD" itemprop="priceCurrency"> <meta content="https://schema.org/InStock" itemprop="availability"> </div> <div class="flex gap-4"> <a class="text-purple-400 hover:text-purple-300" href="https://www.redwoodcity.org/residents/redwood-city-events/music/music-in-the-park" target="_blank" itemprop="url"> Learn More → </a> </div> <meta content="https://schema.org/EventScheduled" itemprop="eventStatus"> <meta content="https://schema.org/OfflineEventAttendanceMode" itemprop="eventAttendanceMode"> </div> <div class="rounded-lg transition-all bg-gray-900/50 border border-purple-500/30 hover:border-purple-400 p-6 ring-2 ring-purple-500" itemscope itemtype="https://schema.org/MusicEvent" itemprop="event"> <meta content="https://sinisterdexter.net#event-2025-09-11-valencia-live!-at-blondie's-bar" itemprop="@id"> <div class="text-xs mb-2 font-bold text-purple-400">FEATURED</div> <div class="font-bold mb-2 text-amber-500"> <time datetime="2025-09-11T5:00 PM-08:00" itemprop="startDate"> THU, SEP 10 </time> </div> <div itemscope itemtype="https://schema.org/Place" itemprop="location"> <h3 class="font-bold mb-2 text-xl" itemprop="name">Valencia LIVE! at Blondie's Bar</h3> <div itemscope itemtype="https://schema.org/PostalAddress" itemprop="address"> <p class="text-gray-400 mb-2"> <span itemprop="addressLocality">San Francisco, CA</span>, <span itemprop="addressRegion">CA</span> </p><meta content="" itemprop="addressCountry"> <p></p> </div> </div> <p class="text-gray-300 mb-4"> <time datetime="2025-09-11T5:00 PM">5:00 PM AM</time> - <time datetime="2025-09-11T8:00 PM" itemprop="endDate">8:00 PM AM</time> </p> <p class="text-sm text-gray-400 mb-4" itemprop="description">Big night market on Valencia St with food, drink, entertainment, and FUNK!</p> <div class="hidden" itemscope itemtype="https://schema.org/MusicGroup" itemprop="performer"> <meta content="Sinister Dexter" itemprop="name"> <link href="https://sinisterdexter.net" itemprop="url"> </div> <div itemscope itemtype="https://schema.org/Offer" itemprop="offers"> <meta content="0" itemprop="price"> <meta content="USD" itemprop="priceCurrency"> <meta content="https://schema.org/InStock" itemprop="availability"> </div> <div class="flex gap-4"> <a class="text-purple-400 hover:text-purple-300" href="https://blondiesbarsf.com/" target="_blank" itemprop="url"> Learn More → </a> </div> <meta content="https://schema.org/EventScheduled" itemprop="eventStatus"> <meta content="https://schema.org/OfflineEventAttendanceMode" itemprop="eventAttendanceMode"> </div> <div class="rounded-lg transition-all bg-gray-900/50 border border-purple-500/30 hover:border-purple-400 p-6" itemscope itemtype="https://schema.org/MusicEvent" itemprop="event"> <meta content="https://sinisterdexter.net#event-2025-11-14-club-fox" itemprop="@id"> <div class="font-bold mb-2 text-amber-500"> <time datetime="2025-11-14T6:00 PM-08:00" itemprop="startDate"> FRI, NOV 13 </time> </div> <div itemscope itemtype="https://schema.org/Place" itemprop="location"> <h3 class="font-bold mb-2 text-xl" itemprop="name">Club Fox</h3> <div itemscope itemtype="https://schema.org/PostalAddress" itemprop="address"> <p class="text-gray-400 mb-2"> <span itemprop="addressLocality">Redwood City, CA</span>, <span itemprop="addressRegion">CA</span> </p><meta content="" itemprop="addressCountry"> <p></p> </div> </div> <p class="text-gray-300 mb-4"> <time datetime="2025-11-14T6:00 PM">6:00 PM AM</time> - <time datetime="2025-11-14T10:00 PM" itemprop="endDate">10:00 PM AM</time> </p> <p class="text-sm text-gray-400 mb-4" itemprop="description">Friday night funk at the legendary Club Fox venue</p> <div class="hidden" itemscope itemtype="https://schema.org/MusicGroup" itemprop="performer"> <meta content="Sinister Dexter" itemprop="name"> <link href="https://sinisterdexter.net" itemprop="url"> </div> <div itemscope itemtype="https://schema.org/Offer" itemprop="offers"> <meta content="0" itemprop="price"> <meta content="USD" itemprop="priceCurrency"> <meta content="https://schema.org/InStock" itemprop="availability"> <meta content="https://www.eventbrite.com/e/sinister-dexter-funk-soul-horns-tickets-1568339939979?aff=ebdssbdestsearch" itemprop="url"> </div> <div class="flex gap-4"> <a class="text-purple-400 hover:text-purple-300" href="https://clubfoxrwc.com/" target="_blank" itemprop="url"> Learn More → </a> <a class="font-bold text-purple-400 hover:text-purple-300" href="https://www.eventbrite.com/e/sinister-dexter-funk-soul-horns-tickets-1568339939979?aff=ebdssbdestsearch" target="_blank"> Buy Tickets → </a> </div> <meta content="https://schema.org/EventScheduled" itemprop="eventStatus"> <meta content="https://schema.org/OfflineEventAttendanceMode" itemprop="eventAttendanceMode"> </div> </div> <div class="mt-16"> <h3 class="text-purple-400 bebas text-3xl mb-8 text-center">PAST SHOWS</h3> <div class="rounded-lg bg-gray-900/30 max-w-4xl mx-auto p-6" id="past-shows-container"> <div class="space-y-3"> <div class="flex justify-between border-b border-gray-800 flex-col last:border-0 py-2 sm:flex-row sm:items-center" itemscope itemtype="https://schema.org/MusicEvent"> <div class="flex-1"> <time datetime="2024-09-29" itemprop="startDate" class="text-sm font-semibold text-amber-500"> SEP 28, 2024 </time> <span class="text-gray-400 mx-2">•</span> <a class="text-gray-300 hover:text-purple-400 transition-colors" href="https://bernalfiesta.com/" target="_blank" itemprop="url"> <span itemprop="name">Fiesta on the Hill</span> <i class="text-xs fa-external-link-alt fas ml-1"></i> </a> <span class="text-sm text-gray-500 ml-2"> <span itemprop="location" itemscope itemtype="https://schema.org/Place"> <span itemprop="address" itemscope itemtype="https://schema.org/PostalAddress"> <span itemprop="addressLocality">Bernal Heights, San Francisco</span> </span> </span> </span> </div> <div class="text-sm text-gray-500 mt-1 sm:mt-0">2:00 PM - 4:00 PM</div> </div> <div class="flex justify-between border-b border-gray-800 flex-col last:border-0 py-2 sm:flex-row sm:items-center" itemscope itemtype="https://schema.org/MusicEvent"> <div class="flex-1"> <time datetime="2024-06-27" itemprop="startDate" class="text-sm font-semibold text-amber-500"> JUN 26, 2024 </time> <span class="text-gray-400 mx-2">•</span> <span class="text-gray-300" itemprop="name">Los Altos Summer Concert Series</span> <span class="text-sm text-gray-500 ml-2"> <span itemprop="location" itemscope itemtype="https://schema.org/Place"> <span itemprop="address" itemscope itemtype="https://schema.org/PostalAddress"> <span itemprop="addressLocality">Grant Park, Los Altos</span> </span> </span> </span> </div> <div class="text-sm text-gray-500 mt-1 sm:mt-0">6:00 PM - 8:00 PM</div> </div> <div class="flex justify-between border-b border-gray-800 flex-col last:border-0 py-2 sm:flex-row sm:items-center" itemscope itemtype="https://schema.org/MusicEvent"> <div class="flex-1"> <time datetime="2024-04-27" itemprop="startDate" class="text-sm font-semibold text-amber-500"> APR 26, 2024 </time> <span class="text-gray-400 mx-2">•</span> <span class="text-gray-300" itemprop="name">Mount Carmel Wild West Days Festival</span> <span class="text-sm text-gray-500 ml-2"> <span itemprop="location" itemscope itemtype="https://schema.org/Place"> <span itemprop="address" itemscope itemtype="https://schema.org/PostalAddress"> <span itemprop="addressLocality">Redwood City</span> </span> </span> </span> </div> <div class="text-sm text-gray-500 mt-1 sm:mt-0">3:00 PM - 5:00 PM</div> </div> <div class="hidden space-y-3" id="hidden-past-shows"> <div class="flex justify-between border-b border-gray-800 flex-col last:border-0 py-2 sm:flex-row sm:items-center" itemscope itemtype="https://schema.org/MusicEvent"> <div class="flex-1"> <time datetime="2024-03-23" itemprop="startDate" class="text-sm font-semibold text-amber-500"> MAR 22, 2024 </time> <span class="text-gray-400 mx-2">•</span> <span class="text-gray-300" itemprop="name">Ireland's 32</span> <span class="text-sm text-gray-500 ml-2"> <span itemprop="location" itemscope itemtype="https://schema.org/Place"> <span itemprop="address" itemscope itemtype="https://schema.org/PostalAddress"> <span itemprop="addressLocality">San Francisco</span> </span> </span> </span> </div> <div class="text-sm text-gray-500 mt-1 sm:mt-0">8:00 PM - 11:00 PM</div> </div> <div class="flex justify-between border-b border-gray-800 flex-col last:border-0 py-2 sm:flex-row sm:items-center" itemscope itemtype="https://schema.org/MusicEvent"> <div class="flex-1"> <time datetime="2024-02-10" itemprop="startDate" class="text-sm font-semibold text-amber-500"> FEB 9, 2024 </time> <span class="text-gray-400 mx-2">•</span> <a class="text-gray-300 hover:text-purple-400 transition-colors" href="https://blueoakbrewing.com" target="_blank" itemprop="url"> <span itemprop="name">Blue Oak Brewing Company</span> <i class="text-xs fa-external-link-alt fas ml-1"></i> </a> <span class="text-sm text-gray-500 ml-2"> <span itemprop="location" itemscope itemtype="https://schema.org/Place"> <span itemprop="address" itemscope itemtype="https://schema.org/PostalAddress"> <span itemprop="addressLocality">San Carlos</span> </span> </span> </span> </div> <div class="text-sm text-gray-500 mt-1 sm:mt-0">3:00 PM - 6:00 PM</div> </div> <div class="flex justify-between border-b border-gray-800 flex-col last:border-0 py-2 sm:flex-row sm:items-center" itemscope itemtype="https://schema.org/MusicEvent"> <div class="flex-1"> <time datetime="2024-02-02" itemprop="startDate" class="text-sm font-semibold text-amber-500"> FEB 1, 2024 </time> <span class="text-gray-400 mx-2">•</span> <a class="text-gray-300 hover:text-purple-400 transition-colors" href="https://www.clubfoxrwc.com" target="_blank" itemprop="url"> <span itemprop="name">Club Fox - 25 Year Reunion Show</span> <i class="text-xs fa-external-link-alt fas ml-1"></i> </a> <span class="text-sm text-gray-500 ml-2"> <span itemprop="location" itemscope itemtype="https://schema.org/Place"> <span itemprop="address" itemscope itemtype="https://schema.org/PostalAddress"> <span itemprop="addressLocality">Redwood City</span> </span> </span> </span> </div> <div class="text-sm text-gray-500 mt-1 sm:mt-0">7:00 PM - 10:00 PM</div> </div> <div class="flex justify-between border-b border-gray-800 flex-col last:border-0 py-2 sm:flex-row sm:items-center" itemscope itemtype="https://schema.org/MusicEvent"> <div class="flex-1"> <time datetime="2023-11-04" itemprop="startDate" class="text-sm font-semibold text-amber-500"> NOV 3, 2023 </time> <span class="text-gray-400 mx-2">•</span> <a class="text-gray-300 hover:text-purple-400 transition-colors" href="https://blueoakbrewing.com" target="_blank" itemprop="url"> <span itemprop="name">Blue Oak Brewing Company</span> <i class="text-xs fa-external-link-alt fas ml-1"></i> </a> <span class="text-sm text-gray-500 ml-2"> <span itemprop="location" itemscope itemtype="https://schema.org/Place"> <span itemprop="address" itemscope itemtype="https://schema.org/PostalAddress"> <span itemprop="addressLocality">San Carlos</span> </span> </span> </span> </div> <div class="text-sm text-gray-500 mt-1 sm:mt-0">3:00 PM - 6:00 PM</div> </div> <div class="flex justify-between border-b border-gray-800 flex-col last:border-0 py-2 sm:flex-row sm:items-center" itemscope itemtype="https://schema.org/MusicEvent"> <div class="flex-1"> <time datetime="2023-10-19" itemprop="startDate" class="text-sm font-semibold text-amber-500"> OCT 18, 2023 </time> <span class="text-gray-400 mx-2">•</span> <a class="text-gray-300 hover:text-purple-400 transition-colors" href="https://blueoakbrewing.com" target="_blank" itemprop="url"> <span itemprop="name">Blue Oak Brewing with TGIFunk</span> <i class="text-xs fa-external-link-alt fas ml-1"></i> </a> <span class="text-sm text-gray-500 ml-2"> <span itemprop="location" itemscope itemtype="https://schema.org/Place"> <span itemprop="address" itemscope itemtype="https://schema.org/PostalAddress"> <span itemprop="addressLocality">San Carlos</span> </span> </span> </span> </div> <div class="text-sm text-gray-500 mt-1 sm:mt-0">3:00 PM - 6:00 PM</div> </div> <div class="flex justify-between border-b border-gray-800 flex-col last:border-0 py-2 sm:flex-row sm:items-center" itemscope itemtype="https://schema.org/MusicEvent"> <div class="flex-1"> <time datetime="2023-09-29" itemprop="startDate" class="text-sm font-semibold text-amber-500"> SEP 28, 2023 </time> <span class="text-gray-400 mx-2">•</span> <span class="text-gray-300" itemprop="name">Fiesta on the Hill</span> <span class="text-sm text-gray-500 ml-2"> <span itemprop="location" itemscope itemtype="https://schema.org/Place"> <span itemprop="address" itemscope itemtype="https://schema.org/PostalAddress"> <span itemprop="addressLocality">Bernal Heights, San Francisco</span> </span> </span> </span> </div> <div class="text-sm text-gray-500 mt-1 sm:mt-0">2:00 PM - 4:00 PM</div> </div> <div class="flex justify-between border-b border-gray-800 flex-col last:border-0 py-2 sm:flex-row sm:items-center" itemscope itemtype="https://schema.org/MusicEvent"> <div class="flex-1"> <time datetime="2023-08-31" itemprop="startDate" class="text-sm font-semibold text-amber-500"> AUG 30, 2023 </time> <span class="text-gray-400 mx-2">•</span> <a class="text-gray-300 hover:text-purple-400 transition-colors" href="https://blueoakbrewing.com" target="_blank" itemprop="url"> <span itemprop="name">Blue Oak Brewing Company</span> <i class="text-xs fa-external-link-alt fas ml-1"></i> </a> <span class="text-sm text-gray-500 ml-2"> <span itemprop="location" itemscope itemtype="https://schema.org/Place"> <span itemprop="address" itemscope itemtype="https://schema.org/PostalAddress"> <span itemprop="addressLocality">San Carlos</span> </span> </span> </span> </div> <div class="text-sm text-gray-500 mt-1 sm:mt-0">3:00 PM - 6:00 PM</div> </div> <div class="flex justify-between border-b border-gray-800 flex-col last:border-0 py-2 sm:flex-row sm:items-center" itemscope itemtype="https://schema.org/MusicEvent"> <div class="flex-1"> <time datetime="2023-08-16" itemprop="startDate" class="text-sm font-semibold text-amber-500"> AUG 15, 2023 </time> <span class="text-gray-400 mx-2">•</span> <span class="text-gray-300" itemprop="name">Redwood City Music in the Park</span> <span class="text-sm text-gray-500 ml-2"> <span itemprop="location" itemscope itemtype="https://schema.org/Place"> <span itemprop="address" itemscope itemtype="https://schema.org/PostalAddress"> <span itemprop="addressLocality">Stafford Park, Redwood City</span> </span> </span> </span> </div> <div class="text-sm text-gray-500 mt-1 sm:mt-0">5:00 PM - 7:00 PM</div> </div> <div class="flex justify-between border-b border-gray-800 flex-col last:border-0 py-2 sm:flex-row sm:items-center" itemscope itemtype="https://schema.org/MusicEvent"> <div class="flex-1"> <time datetime="2023-06-17" itemprop="startDate" class="text-sm font-semibold text-amber-500"> JUN 16, 2023 </time> <span class="text-gray-400 mx-2">•</span> <a class="text-gray-300 hover:text-purple-400 transition-colors" href="https://blueoakbrewing.com" target="_blank" itemprop="url"> <span itemprop="name">Blue Oak Brewing Company</span> <i class="text-xs fa-external-link-alt fas ml-1"></i> </a> <span class="text-sm text-gray-500 ml-2"> <span itemprop="location" itemscope itemtype="https://schema.org/Place"> <span itemprop="address" itemscope itemtype="https://schema.org/PostalAddress"> <span itemprop="addressLocality">San Carlos</span> </span> </span> </span> </div> <div class="text-sm text-gray-500 mt-1 sm:mt-0">3:00 PM - 6:00 PM</div> </div> </div> <button class="text-xs hover:opacity-100 hover:text-gray-500 mt-2 opacity-70 text-gray-600 transition-colors" id="toggle-past-shows" onclick="togglePastShows()"> <span class="text-[10px] tracking-wider uppercase" id="toggle-text">9 more</span> <svg class="inline-block h-2.5 ml-0.5 transition-transform w-2.5" fill="currentColor" viewBox="0 0 20 20" id="toggle-icon"> <path d="M5.293 7.293a1 1 0 011.414 0L10 10.586l3.293-3.293a1 1 0 111.414 1.414l-4 4a1 1 0 01-1.414 0l-4-4a1 1 0 010-1.414z" clip-rule="evenodd" fill-rule="evenodd"/> </svg> </button> </div> </div> </div> </div> </section> <script>function togglePastShows(){const t=document.getElementById("hidden-past-shows"),e=document.getElementById("toggle-text"),n=document.getElementById("toggle-icon");t.classList.contains("hidden")?(t.classList.remove("hidden"),e.textContent="less",n.style.transform="rotate(180deg)"):(t.classList.add("hidden"),e.textContent="9 more",n.style.transform="rotate(0deg)")}</script> <section class="px-6 py-20 bg-gradient-to-b from-black to-purple-950/30" id="music-library" aria-labelledby="music-library-heading"> <div class="mx-auto max-w-4xl"> <h2 class="text-center bebas gradient-text md:text-6xl text-5xl mb-12" id="music-library-heading">Listen to Our Tracks</h2> <div class="rounded-lg bg-gray-900/50 border border-purple-500/30 p-6"> <h3 class="font-bold text-purple-400 mb-4 text-xl">Track Listing</h3> <div class="grid lg:grid-cols-3 md:grid-cols-2 gap-2 grid-cols-1" id="track-list"> <button class="rounded-lg p-3 hover:bg-purple-900/30 text-left track-item transition-all w-full" aria-label="Play Broadband Connection" data-artist="Sinister Dexter" data-duration="3:01" data-src="https://cdn.sinister-dexter.com/music/Sinister_Dexter-01-Broadband_Connection.mp3" data-title="Broadband Connection" data-track-num="1"> <div class="flex items-center justify-between"> <span class="text-gray-300 font-medium">1. Broadband Connection</span> <span class="text-xs text-gray-500">(3:01)</span> </div> </button> <button class="rounded-lg p-3 hover:bg-purple-900/30 text-left track-item transition-all w-full" aria-label="Play Freeman" data-artist="Sinister Dexter" data-duration="3:03" data-src="https://cdn.sinister-dexter.com/music/Sinister_Dexter-02-Freeman.mp3" data-title="Freeman" data-track-num="2"> <div class="flex items-center justify-between"> <span class="text-gray-300 font-medium">2. Freeman</span> <span class="text-xs text-gray-500">(3:03)</span> </div> </button> <button class="rounded-lg p-3 hover:bg-purple-900/30 text-left track-item transition-all w-full" aria-label="Play Chasing the Clouds Away" data-artist="Sinister Dexter" data-duration="3:12" data-src="https://cdn.sinister-dexter.com/music/Sinister_Dexter-02-Chasing_the_Clouds_Away.mp3" data-title="Chasing the Clouds Away" data-track-num="3"> <div class="flex items-center justify-between"> <span class="text-gray-300 font-medium">3. Chasing the Clouds Away</span> <span class="text-xs text-gray-500">(3:12)</span> </div> </button> <button class="rounded-lg p-3 hover:bg-purple-900/30 text-left track-item transition-all w-full" aria-label="Play Morenita" data-artist="Sinister Dexter" data-duration="3:26" data-src="https://cdn.sinister-dexter.com/music/Sinister_Dexter-03-Morenita.mp3" data-title="Morenita" data-track-num="4"> <div class="flex items-center justify-between"> <span class="text-gray-300 font-medium">4. Morenita</span> <span class="text-xs text-gray-500">(3:26)</span> </div> </button> <button class="rounded-lg p-3 hover:bg-purple-900/30 text-left track-item transition-all w-full" aria-label="Play One-Eyed Jack" data-artist="Sinister Dexter" data-duration="4:15" data-src="https://cdn.sinister-dexter.com/music/Sinister_Dexter-03-One-Eyed_Jack.mp3" data-title="One-Eyed Jack" data-track-num="5"> <div class="flex items-center justify-between"> <span class="text-gray-300 font-medium">5. One-Eyed Jack</span> <span class="text-xs text-gray-500">(4:15)</span> </div> </button> <button class="rounded-lg p-3 hover:bg-purple-900/30 text-left track-item transition-all w-full" aria-label="Play Are You Ready to Dance" data-artist="Sinister Dexter" data-duration="3:26" data-src="https://cdn.sinister-dexter.com/music/Sinister_Dexter-04-Are_You_Ready_to_Dance.mp3" data-title="Are You Ready to Dance" data-track-num="6"> <div class="flex items-center justify-between"> <span class="text-gray-300 font-medium">6. Are You Ready to Dance</span> <span class="text-xs text-gray-500">(3:26)</span> </div> </button> <button class="rounded-lg p-3 hover:bg-purple-900/30 text-left track-item transition-all w-full" aria-label="Play Sinister Dexter" data-artist="Sinister Dexter" data-duration="3:26" data-src="https://cdn.sinister-dexter.com/music/Sinister_Dexter-04-Sinister_Dexter.mp3" data-title="Sinister Dexter" data-track-num="7"> <div class="flex items-center justify-between"> <span class="text-gray-300 font-medium">7. Sinister Dexter</span> <span class="text-xs text-gray-500">(3:26)</span> </div> </button> <button class="rounded-lg p-3 hover:bg-purple-900/30 text-left track-item transition-all w-full" aria-label="Play Two Before Nine" data-artist="Sinister Dexter" data-duration="4:05" data-src="https://cdn.sinister-dexter.com/music/Sinister_Dexter-05-Two_Before_Nine.mp3" data-title="Two Before Nine" data-track-num="8"> <div class="flex items-center justify-between"> <span class="text-gray-300 font-medium">8. Two Before Nine</span> <span class="text-xs text-gray-500">(4:05)</span> </div> </button> <button class="rounded-lg p-3 hover:bg-purple-900/30 text-left track-item transition-all w-full" aria-label="Play Can't Take My Eyes Off You" data-artist="Sinister Dexter" data-duration="4:42" data-src="https://cdn.sinister-dexter.com/music/Sinister_Dexter-06-Cant_Take_My_Eyes_Off_You.mp3" data-title="Can't Take My Eyes Off You" data-track-num="9"> <div class="flex items-center justify-between"> <span class="text-gray-300 font-medium">9. Can't Take My Eyes Off You</span> <span class="text-xs text-gray-500">(4:42)</span> </div> </button> <button class="rounded-lg p-3 hover:bg-purple-900/30 text-left track-item transition-all w-full" aria-label="Play Jack in the Box" data-artist="Sinister Dexter" data-duration="2:35" data-src="https://cdn.sinister-dexter.com/music/Sinister_Dexter-09-Jack_in_the_Box.mp3" data-title="Jack in the Box" data-track-num="10"> <div class="flex items-center justify-between"> <span class="text-gray-300 font-medium">10. Jack in the Box</span> <span class="text-xs text-gray-500">(2:35)</span> </div> </button> <button class="rounded-lg p-3 hover:bg-purple-900/30 text-left track-item transition-all w-full" aria-label="Play Him or Me" data-artist="Sinister Dexter" data-duration="4:49" data-src="https://cdn.sinister-dexter.com/music/Sinister_Dexter-10-Him_or_Me.mp3" data-title="Him or Me" data-track-num="11"> <div class="flex items-center justify-between"> <span class="text-gray-300 font-medium">11. Him or Me</span> <span class="text-xs text-gray-500">(4:49)</span> </div> </button> <button class="rounded-lg p-3 hover:bg-purple-900/30 text-left track-item transition-all w-full" aria-label="Play Swim on Over" data-artist="Sinister Dexter" data-duration="4:08" data-src="https://cdn.sinister-dexter.com/music/Sinister_Dexter-11-Swim_on_Over.mp3" data-title="Swim on Over" data-track-num="12"> <div class="flex items-center justify-between"> <span class="text-gray-300 font-medium">12. Swim on Over</span> <span class="text-xs text-gray-500">(4:08)</span> </div> </button> <button class="rounded-lg p-3 hover:bg-purple-900/30 text-left track-item transition-all w-full" aria-label="Play Better Left Unsaid" data-artist="Sinister Dexter" data-duration="4:25" data-src="https://cdn.sinister-dexter.com/music/Sinister_Dexter-12-Better_Left_Unsaid.mp3" data-title="Better Left Unsaid" data-track-num="13"> <div class="flex items-center justify-between"> <span class="text-gray-300 font-medium">13. Better Left Unsaid</span> <span class="text-xs text-gray-500">(4:25)</span> </div> </button> <button class="rounded-lg p-3 hover:bg-purple-900/30 text-left track-item transition-all w-full" aria-label="Play Another Drink" data-artist="Sinister Dexter" data-duration="3:04" data-src="https://cdn.sinister-dexter.com/music/Sinister_Dexter-13-Another_Drink.mp3" data-title="Another Drink" data-track-num="14"> <div class="flex items-center justify-between"> <span class="text-gray-300 font-medium">14. Another Drink</span> <span class="text-xs text-gray-500">(3:04)</span> </div> </button> <button class="rounded-lg p-3 hover:bg-purple-900/30 text-left track-item transition-all w-full" aria-label="Play Show Us How to Get Down (Live)" data-artist="Sinister Dexter" data-duration="5:53" data-src="https://cdn.sinister-dexter.com/music/Sinister_Dexter-01-Show_Us_How_to_Get_Down.mp3" data-title="Show Us How to Get Down (Live)" data-track-num="15"> <div class="flex items-center justify-between"> <span class="text-gray-300 font-medium">15. Show Us How to Get Down (Live)</span> <span class="text-xs text-gray-500">(5:53)</span> </div> </button> <button class="rounded-lg p-3 hover:bg-purple-900/30 text-left track-item transition-all w-full" aria-label="Play Show Us How to Get Down (Studio)" data-artist="Sinister Dexter" data-duration="5:58" data-src="https://cdn.sinister-dexter.com/music/Sinister_Dexter-08-Show_Us_How_to_Get_Down.mp3" data-title="Show Us How to Get Down (Studio)" data-track-num="16"> <div class="flex items-center justify-between"> <span class="text-gray-300 font-medium">16. Show Us How to Get Down (Studio)</span> <span class="text-xs text-gray-500">(5:58)</span> </div> </button> <button class="rounded-lg p-3 hover:bg-purple-900/30 text-left track-item transition-all w-full" aria-label="Play Show Us How to Get Down (Alt)" data-artist="Sinister Dexter" data-duration="6:30" data-src="https://cdn.sinister-dexter.com/music/Sinister_Dexter-Show_Us_How_to_Get_Down.mp3" data-title="Show Us How to Get Down (Alt)" data-track-num="17"> <div class="flex items-center justify-between"> <span class="text-gray-300 font-medium">17. Show Us How to Get Down (Alt)</span> <span class="text-xs text-gray-500">(6:30)</span> </div> </button> <button class="rounded-lg p-3 hover:bg-purple-900/30 text-left track-item transition-all w-full" aria-label="Play Son of a Preacher Man" data-artist="Sinister Dexter" data-duration="3:03" data-src="https://cdn.sinister-dexter.com/music/Sinister_Dexter-Son_of_a_Preacher_Man.mp3" data-title="Son of a Preacher Man" data-track-num="18"> <div class="flex items-center justify-between"> <span class="text-gray-300 font-medium">18. Son of a Preacher Man</span> <span class="text-xs text-gray-500">(3:03)</span> </div> </button> <button class="rounded-lg p-3 hover:bg-purple-900/30 text-left track-item transition-all w-full" aria-label="Play Dream Come True" data-artist="Sinister Dexter" data-duration="5:21" data-src="https://cdn.sinister-dexter.com/music/Sinister_Dexter-Dream_Come_True.mp3" data-title="Dream Come True" data-track-num="19"> <div class="flex items-center justify-between"> <span class="text-gray-300 font-medium">19. Dream Come True</span> <span class="text-xs text-gray-500">(5:21)</span> </div> </button> <button class="rounded-lg p-3 hover:bg-purple-900/30 text-left track-item transition-all w-full" aria-label="Play What is Hip" data-artist="Sinister Dexter" data-duration="5:51" data-src="https://cdn.sinister-dexter.com/music/Sinister_Dexter-What_is_Hip.mp3" data-title="What is Hip" data-track-num="20"> <div class="flex items-center justify-between"> <span class="text-gray-300 font-medium">20. What is Hip</span> <span class="text-xs text-gray-500">(5:51)</span> </div> </button> <button class="rounded-lg p-3 hover:bg-purple-900/30 text-left track-item transition-all w-full" aria-label="Play The Letter" data-artist="Sinister Dexter" data-duration="3:58" data-src="https://cdn.sinister-dexter.com/music/Sinister_Dexter-The_Letter.mp3" data-title="The Letter" data-track-num="21"> <div class="flex items-center justify-between"> <span class="text-gray-300 font-medium">21. The Letter</span> <span class="text-xs text-gray-500">(3:58)</span> </div> </button> </div> <div class="border-gray-700 border-t mt-6 pt-6"> <p class="text-sm text-gray-400 text-center"> All original songs licensed under <a class="text-purple-400 hover:text-purple-300 underline" href="https://creativecommons.org/licenses/by-nc-sa/2.0/" target="_blank" rel="noopener noreferrer">Creative Commons BY-NC-SA 2.0</a> </p> </div> </div> </div> </section> <section class="px-6 py-20 bg-gradient-to-b from-purple-950/30 to-black" id="videos" aria-labelledby="videos-heading" itemscope itemtype="https://schema.org/ItemList"> <div class="mx-auto max-w-7xl"> <h2 class="text-center bebas gradient-text md:text-6xl text-5xl mb-12" id="videos-heading">Live Performances</h2> <div class="grid lg:grid-cols-3 md:grid-cols-2 gap-8"> <div class="rounded-lg transition-all bg-gray-900/50 border border-purple-500/30 hover:border-purple-400 overflow-hidden" itemscope itemtype="https://schema.org/VideoObject" itemprop="itemListElement"> <a class="group block" href="https://www.youtube.com/watch?v=D9gIdE7IJsY" target="_blank" aria-label="Watch Son of a Preacher Man on YouTube" rel="noopener noreferrer"> <div class="relative overflow-hidden aspect-video"> <img alt="Son of a Preacher Man - Live performance thumbnail" class="transition-transform duration-300 group-hover:scale-105 h-full object-cover w-full" src="https://img.youtube.com/vi/D9gIdE7IJsY/maxresdefault.jpg" loading="lazy"> <div class="flex items-center justify-center absolute bg-black/30 group-hover:bg-black/40 inset-0 transition-colors"> <div class="rounded-lg bg-red-600 group-hover:scale-110 p-4 transition-transform"> <svg class="w-8 h-8 text-white" fill="currentColor" viewBox="0 0 24 24"> <path d="M8 5v14l11-7z"/> </svg> </div> </div> </div> </a> <div class="p-4"> <h3 class="text-purple-400 text-lg font-semibold" itemprop="name">Son of a Preacher Man</h3> <p class="text-sm text-gray-400" itemprop="description">Redwood City Music In The Park, 2012</p> <meta content="2012-08-15" itemprop="uploadDate"> <meta content="https://img.youtube.com/vi/D9gIdE7IJsY/maxresdefault.jpg" itemprop="thumbnailUrl"> <meta content="https://www.youtube.com/watch?v=D9gIdE7IJsY" itemprop="contentUrl"> <meta content="https://www.youtube.com/embed/D9gIdE7IJsY" itemprop="embedUrl"> <meta content="PT3M30S" itemprop="duration"> </div> </div> <div class="rounded-lg transition-all bg-gray-900/50 border border-purple-500/30 hover:border-purple-400 overflow-hidden" itemscope itemtype="https://schema.org/VideoObject" itemprop="itemListElement"> <a class="group block" href="https://www.youtube.com/watch?v=vWRJUqNdnX0" target="_blank" aria-label="Watch I Can't Stand The Rain on YouTube" rel="noopener noreferrer"> <div class="relative overflow-hidden aspect-video"> <img alt="I Can't Stand The Rain - Live performance thumbnail" class="transition-transform duration-300 group-hover:scale-105 h-full object-cover w-full" src="https://img.youtube.com/vi/vWRJUqNdnX0/maxresdefault.jpg" loading="lazy"> <div class="flex items-center justify-center absolute bg-black/30 group-hover:bg-black/40 inset-0 transition-colors"> <div class="rounded-lg bg-red-600 group-hover:scale-110 p-4 transition-transform"> <svg class="w-8 h-8 text-white" fill="currentColor" viewBox="0 0 24 24"> <path d="M8 5v14l11-7z"/> </svg> </div> </div> </div> </a> <div class="p-4"> <h3 class="text-purple-400 text-lg font-semibold" itemprop="name">I Can't Stand The Rain</h3> <p class="text-sm text-gray-400" itemprop="description">Redwood City Music In The Park, 2012</p> <meta content="2012-08-15" itemprop="uploadDate"> <meta content="https://img.youtube.com/vi/vWRJUqNdnX0/maxresdefault.jpg" itemprop="thumbnailUrl"> <meta content="https://www.youtube.com/watch?v=vWRJUqNdnX0" itemprop="contentUrl"> <meta content="https://www.youtube.com/embed/vWRJUqNdnX0" itemprop="embedUrl"> <meta content="PT4M15S" itemprop="duration"> </div> </div> <div class="rounded-lg transition-all bg-gray-900/50 border border-purple-500/30 hover:border-purple-400 overflow-hidden" itemscope itemtype="https://schema.org/VideoObject" itemprop="itemListElement"> <a class="group block" href="https://www.youtube.com/watch?v=buwFWtEuOyQ" target="_blank" aria-label="Watch Broadband Connection on YouTube" rel="noopener noreferrer"> <div class="relative overflow-hidden aspect-video"> <img alt="Broadband Connection - Live performance thumbnail" class="transition-transform duration-300 group-hover:scale-105 h-full object-cover w-full" src="https://img.youtube.com/vi/buwFWtEuOyQ/maxresdefault.jpg" loading="lazy"> <div class="flex items-center justify-center absolute bg-black/30 group-hover:bg-black/40 inset-0 transition-colors"> <div class="rounded-lg bg-red-600 group-hover:scale-110 p-4 transition-transform"> <svg class="w-8 h-8 text-white" fill="currentColor" viewBox="0 0 24 24"> <path d="M8 5v14l11-7z"/> </svg> </div> </div> </div> </a> <div class="p-4"> <h3 class="text-purple-400 text-lg font-semibold" itemprop="name">Broadband Connection</h3> <p class="text-sm text-gray-400" itemprop="description">Redwood City Music In The Park, 2012</p> <meta content="2012-08-15" itemprop="uploadDate"> <meta content="https://img.youtube.com/vi/buwFWtEuOyQ/maxresdefault.jpg" itemprop="thumbnailUrl"> <meta content="https://www.youtube.com/watch?v=buwFWtEuOyQ" itemprop="contentUrl"> <meta content="https://www.youtube.com/embed/buwFWtEuOyQ" itemprop="embedUrl"> <meta content="PT3M01S" itemprop="duration"> </div> </div> </div> <div class="text-center mt-12"> <a class="transition-all rounded-full hover:scale-105 py-3 items-center px-6 inline-flex font-semibold bg-red-600 gap-2 hover:bg-red-700" href="https://www.youtube.com/@SinDexBand" target="_blank" rel="noopener noreferrer"> <svg class="h-5 w-5" fill="currentColor" viewBox="0 0 24 24"> <path d="M23.498 6.186a3.016 3.016 0 0 0-2.122-2.136C19.505 3.545 12 3.545 12 3.545s-7.505 0-9.377.505A3.017 3.017 0 0 0 .502 6.186C0 8.07 0 12 0 12s0 3.93.502 5.814a3.016 3.016 0 0 0 2.122 2.136c1.871.505 9.376.505 9.376.505s7.505 0 9.377-.505a3.015 3.015 0 0 0 2.122-2.136C24 15.93 24 12 24 12s0-3.93-.502-5.814zM9.545 15.568V8.432L15.818 12l-6.273 3.568z"/> </svg> Watch More on YouTube </a> </div> </div> </section> <section class="px-6 py-20 bg-gradient-to-b from-black to-purple-950/30" id="gallery" aria-labelledby="gallery-heading"> <div class="mx-auto max-w-7xl"> <h2 class="text-center bebas gradient-text md:text-6xl text-5xl mb-12" id="gallery-heading">Gallery</h2> <div class="gap-4 grid grid-cols-2 lg:grid-cols-4 md:grid-cols-3" id="gallery-container"> <div class="rounded-lg overflow-hidden aspect-square bg-gray-800 cursor-pointer gallery-item relative"> <picture> <source srcset="/images/thumbnail/IMG_2775_converted.webp" type="image/webp"> <img alt="Sinister Dexter performance photo 0" class="h-full object-cover w-full" src="/images/thumbnail/IMG_2775_converted.jpg" loading="lazy" data-large-jpg="/images/large/IMG_2775_converted.jpg" data-large-webp="/images/large/IMG_2775_converted.webp"> </picture> </div> <div class="rounded-lg overflow-hidden aspect-square bg-gray-800 cursor-pointer gallery-item relative"> <picture> <source srcset="/images/thumbnail/IMG_2795_converted.webp" type="image/webp"> <img alt="Sinister Dexter performance photo 1" class="h-full object-cover w-full" src="/images/thumbnail/IMG_2795_converted.jpg" loading="lazy" data-large-jpg="/images/large/IMG_2795_converted.jpg" data-large-webp="/images/large/IMG_2795_converted.webp"> </picture> </div> <div class="rounded-lg overflow-hidden aspect-square bg-gray-800 cursor-pointer gallery-item relative"> <picture> <source srcset="/images/thumbnail/sindex-full.webp" type="image/webp"> <img alt="Sinister Dexter performance photo 2" class="h-full object-cover w-full" src="/images/thumbnail/sindex-full.jpg" loading="lazy" data-large-jpg="/images/large/sindex-full.jpg" data-large-webp="/images/large/sindex-full.webp"> </picture> </div> <div class="rounded-lg overflow-hidden aspect-square bg-gray-800 cursor-pointer gallery-item relative"> <picture> <source srcset="/images/thumbnail/sindex1.webp" type="image/webp"> <img alt="Sinister Dexter performance photo 3" class="h-full object-cover w-full" src="/images/thumbnail/sindex1.jpg" loading="lazy" data-large-jpg="/images/large/sindex1.jpg" data-large-webp="/images/large/sindex1.webp"> </picture> </div> <div class="rounded-lg overflow-hidden aspect-square bg-gray-800 cursor-pointer gallery-item relative"> <picture> <source srcset="/images/thumbnail/sindex2_converted.webp" type="image/webp"> <img alt="Sinister Dexter performance photo 4" class="h-full object-cover w-full" src="/images/thumbnail/sindex2_converted.jpg" loading="lazy" data-large-jpg="/images/large/sindex2_converted.jpg" data-large-webp="/images/large/sindex2_converted.webp"> </picture> </div> <div class="rounded-lg overflow-hidden aspect-square bg-gray-800 cursor-pointer gallery-item relative"> <picture> <source srcset="/images/thumbnail/sindex3.webp" type="image/webp"> <img alt="Sinister Dexter performance photo 5" class="h-full object-cover w-full" src="/images/thumbnail/sindex3.jpg" loading="lazy" data-large-jpg="/images/large/sindex3.jpg" data-large-webp="/images/large/sindex3.webp"> </picture> </div> </div> </div> </section> <div class="flex items-center justify-center inset-0 bg-black/95 fixed hidden p-4 z-[100]" id="lightbox"> <button class="hover:text-gray-300 absolute right-4 text-4xl text-white top-4 z-[101]" aria-label="Close lightbox" id="lightbox-close"> × </button> <img alt="" class="max-h-full max-w-full object-contain" src="" id="lightbox-image"> </div> <script>document.addEventListener("DOMContentLoaded",function(){const e=document.getElementById("lightbox"),t=document.getElementById("lightbox-image"),n=document.getElementById("lightbox-close");function d(){e.classList.add("hidden"),document.body.style.overflow=""}document.querySelectorAll(".gallery-item").forEach(n=>{n.addEventListener("click",function(){const n=this.querySelector("img"),d=n.dataset.largeWebp,c=n.dataset.largeJpg,o=document.createElement("picture");if(d){const e=document.createElement("source");e.srcset=d,e.type="image/webp",o.appendChild(e)}t.src=c||n.src,t.alt=n.alt,e.classList.remove("hidden"),document.body.style.overflow="hidden"})}),n?.addEventListener("click",d),e?.addEventListener("click",function(t){t.target===e&&d()}),document.addEventListener("keydown",function(t){"Escape"!==t.key||e.classList.contains("hidden")||d()})})</script> <section class="px-6 py-20 bg-gradient-to-b from-purple-950/30 to-black" id="about" aria-labelledby="about-heading"> <div class="mx-auto max-w-4xl"> <h2 class="text-center bebas gradient-text md:text-6xl text-5xl mb-12" id="about-heading">About the Band</h2> <div class="rounded-lg bg-gray-900/50 border border-purple-500/30 p-8 mb-8"> <h3 class="text-purple-400 bebas text-3xl mb-4">Over 25 Years of Funk</h3> <p class="text-lg mb-6" itemprop="foundingDate" content="1998"> Formed in the summer of 1998 by members of the Stanford Band, <span class="font-bold text-purple-400" itemprop="alternateName">Sinister Dexter</span> was created to fill the need for a great local party band. Started by horn players, we put together a band that we wanted to go see – which of course meant a horn-heavy sound and classic funk music inspired by Tower of Power and the Blues Brothers. </p> <p class="text-lg mb-6"> We started by playing college parties and small Bay Area clubs like the Quarter Note in Sunnyvale and Blue Chalk Cafe in Palo Alto. In May 2004, we got our first chance to play in San Francisco at the legendary Lost & Found Saloon, where we became regulars and played a sold-out New Year's Eve show to close out 2004. </p> <p class="text-lg mb-6"> As the band matured, we expanded into blues, rock, ska, and Latin jazz – anything with a solid groove that we could put great horn parts to. In 2004, we recorded and self-produced our debut CD, <span class="text-amber-400 italic">Dexterity</span>, featuring 16 original tracks spanning all the genres we'd explored over the years. </p> <p class="text-lg mb-6"> In mid-2005, lightning struck when we recruited the incredible <span class="font-bold text-amber-400" itemprop="member" itemscope itemtype="https://schema.org/Person"><span itemprop="name">Rebecca Lipon</span></span> as our new lead vocalist. A founding member of the jazz quartet Off Beats with over 15 years of opera and musical theater training, Rebecca forced us to take things to the next level. We refined our sound, trimmed the fat, and got back to our core concept as a funk and soul band. </p> <p class="text-lg"> Today, we're an 11-piece soul machine featuring Rebecca's powerhouse vocals, a hot six-piece horn section, and a rock-solid rhythm section. Our high-energy shows have been described as "Tower of Power, Ozomatli, and Joss Stone in a blender" – bringing <span class="text-purple-400 italic">the right to be funky</span> to venues throughout the Bay Area. </p> <meta content="Funk" itemprop="genre"> <meta content="Soul" itemprop="genre"> <meta content="R&B" itemprop="genre"> <meta content="Blues" itemprop="genre"> <div itemscope itemtype="https://schema.org/Place" itemprop="foundingLocation"> <meta content="San Francisco Bay Area" itemprop="name"> <div itemscope itemtype="https://schema.org/PostalAddress" itemprop="address"> <meta content="CA" itemprop="addressRegion"> <meta content="US" itemprop="addressCountry"> </div> </div> </div> </div> </section> <section class="px-6 py-20 bg-gradient-to-r from-purple-900/30 to-amber-900/30" id="pull-quote"> <div class="text-center mx-auto max-w-4xl"> <blockquote class="relative"> <svg class="-left-4 -top-8 absolute h-16 text-purple-500/30 w-16" fill="currentColor" viewBox="0 0 32 32"> <path d="M9.352 4C4.456 7.456 1 13.12 1 19.36c0 5.088 3.072 8.064 6.624 8.064 3.36 0 5.856-2.688 5.856-5.856 0-3.168-2.208-5.472-5.088-5.472-.576 0-1.344.096-1.536.192.48-3.264 3.552-7.104 6.624-9.024L9.352 4zm16.512 0c-4.8 3.456-8.256 9.12-8.256 15.36 0 5.088 3.072 8.064 6.624 8.064 3.264 0 5.856-2.688 5.856-5.856 0-3.168-2.304-5.472-5.184-5.472-.576 0-1.248.096-1.44.192.48-3.264 3.456-7.104 6.528-9.024L25.864 4z"/> </svg> <p class="mb-6 text-2xl font-light italic md:text-3xl text-white"> Tower of Power, Ozomatli, and Joss Stone in a blender </p> <cite class="text-purple-400 text-lg not-italic">— Mountain View Voice</cite> </blockquote> </div> </section> <section class="px-6 py-20 bg-gradient-to-b from-purple-950/30 to-black" id="band-members" aria-labelledby="members-heading"> <div class="mx-auto max-w-7xl"> <h2 class="text-center bebas gradient-text md:text-6xl text-5xl mb-4" id="members-heading">The Band</h2> <p class="text-gray-400 text-center max-w-3xl mb-12 mx-auto"> 11 talented musicians bringing you the finest funk and soul </p> <div class="grid gap-6 grid-cols-2 md:grid-cols-3 lg:grid-cols-4"> <div class="text-center group" itemscope itemtype="https://schema.org/Person" itemprop="member"> <meta content="https://sinisterdexter.net#rebecca-lipon-weekly" itemprop="@id"> <meta content="Sinister Dexter" itemprop="affiliation"> <div class="relative inline-block mb-3"> <picture> <source srcset="images/members/rebecca.webp" type="image/webp"> <img alt="Rebecca Lipon Weekly" class="transition-all rounded-full border-2 border-purple-500 group-hover:border-purple-400 group-hover:scale-105 h-32 mx-auto object-cover w-32" src="images/members/rebecca.jpg" itemprop="image"> </picture> </div> <h3 class="font-bold text-white" itemprop="name">Rebecca Lipon Weekly</h3> <p class="text-sm text-purple-400"> <span class="inline-block mr-1">🎤</span> <span itemprop="jobTitle">Lead Vocals</span> </p> <div class="hidden"> <p itemprop="description">Rebecca's love of singing started with Sesame Street, much to her siblings' chagrin, and it never stopped. She was a founding member of the Off Beats, a Jazz quartet, in her late teens, studied opera and musical theater styles for over 15 years, and performed professionally with theater groups in Northern California, Boston and New York. She has been Sinister Dexter's lead singer since 2005, and loves the mix of soul, funk and blues music.</p> <meta content="Sinister Dexter (2005-present)" itemprop="memberOf"> </div> <div class="rounded-lg p-3 -translate-x-1/2 absolute bg-gray-900 group-hover:opacity-100 left-1/2 mt-2 opacity-0 pointer-events-none shadow-xl transform transition-opacity w-64 z-20"> <p class="text-xs mb-2 text-gray-300">Rebecca's love of singing started with Sesame Street, much to her siblings' chagrin, and it never st...</p> <p class="text-xs text-purple-400">Since 2005</p> </div> </div> <div class="text-center group" itemscope itemtype="https://schema.org/Person" itemprop="member"> <meta content="https://sinisterdexter.net#daniel-walling" itemprop="@id"> <meta content="Sinister Dexter" itemprop="affiliation"> <div class="relative inline-block mb-3"> <picture> <source srcset="images/members/walling.webp" type="image/webp"> <img alt="Daniel Walling" class="transition-all rounded-full border-2 border-purple-500 group-hover:border-purple-400 group-hover:scale-105 h-32 mx-auto object-cover w-32" src="images/members/walling.jpg" itemprop="image"> </picture> </div> <h3 class="font-bold text-white" itemprop="name">Daniel Walling</h3> <p class="text-sm text-purple-400"> <span class="inline-block mr-1">🎹</span> <span itemprop="jobTitle">Keyboard</span> </p> <div class="hidden"> <p itemprop="description">Daniel started playing piano like everyone else, because Mom made him take lessons. But then he got hooked, switching from classical to jazz in high school, playing in various jazz combos in college, and in 2004 switching again to funky tasty groovaliciousness with sindex. Critics have described his organ solos as "really f*#$ing loud".</p> <meta content="Sinister Dexter (2004-present)" itemprop="memberOf"> </div> <div class="rounded-lg p-3 -translate-x-1/2 absolute bg-gray-900 group-hover:opacity-100 left-1/2 mt-2 opacity-0 pointer-events-none shadow-xl transform transition-opacity w-64 z-20"> <p class="text-xs mb-2 text-gray-300">Daniel started playing piano like everyone else, because Mom made him take lessons. But then he got ...</p> <p class="text-xs text-purple-400">Since 2004</p> </div> </div> <div class="text-center group" itemscope itemtype="https://schema.org/Person" itemprop="member"> <meta content="https://sinisterdexter.net#david-mitby" itemprop="@id"> <meta content="Sinister Dexter" itemprop="affiliation"> <div class="relative inline-block mb-3"> <picture> <source srcset="images/members/mitby.webp" type="image/webp"> <img alt="David Mitby" class="transition-all rounded-full border-2 border-purple-500 group-hover:border-purple-400 group-hover:scale-105 h-32 mx-auto object-cover w-32" src="images/members/mitby.jpg" itemprop="image"> </picture> </div> <h3 class="font-bold text-white" itemprop="name">David Mitby</h3> <p class="text-sm text-purple-400"> <span class="inline-block mr-1">🎸</span> <span itemprop="jobTitle">Bass</span> </p> <div class="hidden"> <p itemprop="description">Keeping the groove tight</p> </div> <div class="rounded-lg p-3 -translate-x-1/2 absolute bg-gray-900 group-hover:opacity-100 left-1/2 mt-2 opacity-0 pointer-events-none shadow-xl transform transition-opacity w-64 z-20"> <p class="text-xs mb-2 text-gray-300">Keeping the groove tight</p> </div> </div> <div class="text-center group" itemscope itemtype="https://schema.org/Person" itemprop="member"> <meta content="https://sinisterdexter.net#matt-spitz" itemprop="@id"> <meta content="Sinister Dexter" itemprop="affiliation"> <div class="relative inline-block mb-3"> <picture> <source srcset="images/members/spitz.webp" type="image/webp"> <img alt="Matt Spitz" class="transition-all rounded-full border-2 border-purple-500 group-hover:border-purple-400 group-hover:scale-105 h-32 mx-auto object-cover w-32" src="images/members/spitz.jpg" itemprop="image"> </picture> </div> <h3 class="font-bold text-white" itemprop="name">Matt Spitz</h3> <p class="text-sm text-purple-400"> <span class="inline-block mr-1">🎸</span> <span itemprop="jobTitle">Guitar</span> </p> <div class="hidden"> <p itemprop="description">Matt plays guitar, has performed in 38 states, and recently won a Grammy. As of this writing, at least one of those statements is correct. This is his second tour of duty with SinDex after a decade hiatus in New York City.</p> </div> <div class="rounded-lg p-3 -translate-x-1/2 absolute bg-gray-900 group-hover:opacity-100 left-1/2 mt-2 opacity-0 pointer-events-none shadow-xl transform transition-opacity w-64 z-20"> <p class="text-xs mb-2 text-gray-300">Matt plays guitar, has performed in 38 states, and recently won a Grammy. As of this writing, at lea...</p> </div> </div> <div class="text-center group" itemscope itemtype="https://schema.org/Person" itemprop="member"> <meta content="https://sinisterdexter.net#jon-wagner" itemprop="@id"> <meta content="Sinister Dexter" itemprop="affiliation"> <div class="relative inline-block mb-3"> <picture> <source srcset="images/members/wagner.webp" type="image/webp"> <img alt="Jon Wagner" class="transition-all rounded-full border-2 border-purple-500 group-hover:border-purple-400 group-hover:scale-105 h-32 mx-auto object-cover w-32" src="images/members/wagner.jpg" itemprop="image"> </picture> <div class="text-sm absolute bg-purple-600 border-2 border-white cursor-help flex font-bold group/badge h-8 items-center justify-center right-0 rounded-full shadow-md text-white top-0 w-8"> <span class="pointer-events-none">F</span> <div class="text-xs hidden -translate-x-1/2 absolute bg-gray-900 group-hover/badge:block left-1/2 mt-1 pointer-events-none px-2 py-1 rounded shadow-lg text-white top-full transform whitespace-nowrap z-10"> Founding member (1998-present) </div> </div> </div> <h3 class="font-bold text-white" itemprop="name">Jon Wagner</h3> <p class="text-sm text-purple-400"> <span class="inline-block mr-1">🥁</span> <span itemprop="jobTitle">Drums</span> </p> <div class="hidden"> <p itemprop="description"></p> </div> <div class="rounded-lg p-3 -translate-x-1/2 absolute bg-gray-900 group-hover:opacity-100 left-1/2 mt-2 opacity-0 pointer-events-none shadow-xl transform transition-opacity w-64 z-20"> <p class="text-xs mb-2 text-gray-300"></p> </div> </div> <div class="text-center group" itemscope itemtype="https://schema.org/Person" itemprop="member"> <meta content="https://sinisterdexter.net#kevin-wooley" itemprop="@id"> <meta content="Sinister Dexter" itemprop="affiliation"> <div class="relative inline-block mb-3"> <picture> <source srcset="images/members/wooley.webp" type="image/webp"> <img alt="Kevin Wooley" class="transition-all rounded-full border-2 border-purple-500 group-hover:border-purple-400 group-hover:scale-105 h-32 mx-auto object-cover w-32" src="images/members/wooley.jpg" itemprop="image"> </picture> <div class="text-sm absolute bg-purple-600 border-2 border-white cursor-help flex font-bold group/badge h-8 items-center justify-center right-0 rounded-full shadow-md text-white top-0 w-8"> <span class="pointer-events-none">F</span> <div class="text-xs hidden -translate-x-1/2 absolute bg-gray-900 group-hover/badge:block left-1/2 mt-1 pointer-events-none px-2 py-1 rounded shadow-lg text-white top-full transform whitespace-nowrap z-10"> Founding member (1998-present) </div> </div> </div> <h3 class="font-bold text-white" itemprop="name">Kevin Wooley</h3> <p class="text-sm text-purple-400"> <span class="inline-block mr-1">🎷</span> <span itemprop="jobTitle">Tenor Sax</span> </p> <div class="hidden"> <p itemprop="description">Kevin is a founding member of Sinister Dexter and the defacto manager. He's been playing tenor sax since he was 12 years old in jazz groups, symphonic orchestras, and marching bands. He also plays clarinet and soprano sax, but Kenny G ruined that instrument for everyone. Because of his love for music from the sixties and seventies (and disgust with music from the eighties and nineties), he often thinks he was born a couple of decades too late. By day, Kevin is an Oscar-winning engineer for Industrial Light & Magic.</p> </div> <div class="rounded-lg p-3 -translate-x-1/2 absolute bg-gray-900 group-hover:opacity-100 left-1/2 mt-2 opacity-0 pointer-events-none shadow-xl transform transition-opacity w-64 z-20"> <p class="text-xs mb-2 text-gray-300">Kevin is a founding member of Sinister Dexter and the defacto manager. He's been playing tenor sax s...</p> </div> </div> <div class="text-center group" itemscope itemtype="https://schema.org/Person" itemprop="member"> <meta content="https://sinisterdexter.net#don-hoffman" itemprop="@id"> <meta content="Sinister Dexter" itemprop="affiliation"> <div class="relative inline-block mb-3"> <picture> <source srcset="images/members/hoffman.webp" type="image/webp"> <img alt="Don Hoffman" class="transition-all rounded-full border-2 border-purple-500 group-hover:border-purple-400 group-hover:scale-105 h-32 mx-auto object-cover w-32" src="images/members/hoffman.jpg" itemprop="image"> </picture> <div class="text-sm absolute bg-purple-600 border-2 border-white cursor-help flex font-bold group/badge h-8 items-center justify-center right-0 rounded-full shadow-md text-white top-0 w-8"> <span class="pointer-events-none">F</span> <div class="text-xs hidden -translate-x-1/2 absolute bg-gray-900 group-hover/badge:block left-1/2 mt-1 pointer-events-none px-2 py-1 rounded shadow-lg text-white top-full transform whitespace-nowrap z-10"> Founding member (1998-present) </div> </div> </div> <h3 class="font-bold text-white" itemprop="name">Don Hoffman</h3> <p class="text-sm text-purple-400"> <span class="inline-block mr-1">🎺</span> <span itemprop="jobTitle">Lead Trombone</span> </p> <div class="hidden"> <p itemprop="description">Don's been playing for over 20 years, and his background covers everything from jazz to musical pit orchestras, and symphony to marching bands. He's one of the founding members of Sinister Dexter, and enjoys occasionally playing with other Bay Area bands including Foreverland, which he also helped start.</p> </div> <div class="rounded-lg p-3 -translate-x-1/2 absolute bg-gray-900 group-hover:opacity-100 left-1/2 mt-2 opacity-0 pointer-events-none shadow-xl transform transition-opacity w-64 z-20"> <p class="text-xs mb-2 text-gray-300">Don's been playing for over 20 years, and his background covers everything from jazz to musical pit ...</p> </div> </div> <div class="text-center group" itemscope itemtype="https://schema.org/Person" itemprop="member"> <meta content="https://sinisterdexter.net#josh-schiller" itemprop="@id"> <meta content="Sinister Dexter" itemprop="affiliation"> <div class="relative inline-block mb-3"> <picture> <source srcset="images/members/schiller.webp" type="image/webp"> <img alt="Josh Schiller" class="transition-all rounded-full border-2 border-purple-500 group-hover:border-purple-400 group-hover:scale-105 h-32 mx-auto object-cover w-32" src="images/members/schiller.jpg" itemprop="image"> </picture> <div class="text-sm absolute bg-purple-600 border-2 border-white cursor-help flex font-bold group/badge h-8 items-center justify-center right-0 rounded-full shadow-md text-white top-0 w-8"> <span class="pointer-events-none">F</span> <div class="text-xs hidden -translate-x-1/2 absolute bg-gray-900 group-hover/badge:block left-1/2 mt-1 pointer-events-none px-2 py-1 rounded shadow-lg text-white top-full transform whitespace-nowrap z-10"> Founding member (1998-present) </div> </div> </div> <h3 class="font-bold text-white" itemprop="name">Josh Schiller</h3> <p class="text-sm text-purple-400"> <span class="inline-block mr-1">🎺</span> <span itemprop="jobTitle">Bass Trombone</span> </p> <div class="hidden"> <p itemprop="description">Josh has been playing bone for a very long time. You could even say it pays him more than anything else he does, especially since he's currently enjoying a life of leisure.</p> </div> <div class="rounded-lg p-3 -translate-x-1/2 absolute bg-gray-900 group-hover:opacity-100 left-1/2 mt-2 opacity-0 pointer-events-none shadow-xl transform transition-opacity w-64 z-20"> <p class="text-xs mb-2 text-gray-300">Josh has been playing bone for a very long time. You could even say it pays him more than anything e...</p> </div> </div> <div class="text-center group" itemscope itemtype="https://schema.org/Person" itemprop="member"> <meta content="https://sinisterdexter.net#patrick-neschleba" itemprop="@id"> <meta content="Sinister Dexter" itemprop="affiliation"> <div class="relative inline-block mb-3"> <picture> <source srcset="images/members/patrick.webp" type="image/webp"> <img alt="Patrick Neschleba" class="transition-all rounded-full border-2 border-purple-500 group-hover:border-purple-400 group-hover:scale-105 h-32 mx-auto object-cover w-32" src="images/members/patrick.jpg" itemprop="image"> </picture> <div class="text-sm absolute bg-purple-600 border-2 border-white cursor-help flex font-bold group/badge h-8 items-center justify-center right-0 rounded-full shadow-md text-white top-0 w-8"> <span class="pointer-events-none">F</span> <div class="text-xs hidden -translate-x-1/2 absolute bg-gray-900 group-hover/badge:block left-1/2 mt-1 pointer-events-none px-2 py-1 rounded shadow-lg text-white top-full transform whitespace-nowrap z-10"> Founding member (1998-present) </div> </div> </div> <h3 class="font-bold text-white" itemprop="name">Patrick Neschleba</h3> <p class="text-sm text-purple-400"> <span class="inline-block mr-1">🎷</span> <span itemprop="jobTitle">Alto & Bari Sax</span> </p> <div class="hidden"> <p itemprop="description"></p> </div> <div class="rounded-lg p-3 -translate-x-1/2 absolute bg-gray-900 group-hover:opacity-100 left-1/2 mt-2 opacity-0 pointer-events-none shadow-xl transform transition-opacity w-64 z-20"> <p class="text-xs mb-2 text-gray-300"></p> </div> </div> <div class="text-center group" itemscope itemtype="https://schema.org/Person" itemprop="member"> <meta content="https://sinisterdexter.net#adam-elman" itemprop="@id"> <meta content="Sinister Dexter" itemprop="affiliation"> <div class="relative inline-block mb-3"> <picture> <source srcset="images/members/elman.webp" type="image/webp"> <img alt="Adam Elman" class="transition-all rounded-full border-2 border-purple-500 group-hover:border-purple-400 group-hover:scale-105 h-32 mx-auto object-cover w-32" src="images/members/elman.jpg" itemprop="image"> </picture> <div class="text-sm absolute bg-purple-600 border-2 border-white cursor-help flex font-bold group/badge h-8 items-center justify-center right-0 rounded-full shadow-md text-white top-0 w-8"> <span class="pointer-events-none">F</span> <div class="text-xs hidden -translate-x-1/2 absolute bg-gray-900 group-hover/badge:block left-1/2 mt-1 pointer-events-none px-2 py-1 rounded shadow-lg text-white top-full transform whitespace-nowrap z-10"> Founding member (1998-present) </div> </div> </div> <h3 class="font-bold text-white" itemprop="name">Adam Elman</h3> <p class="text-sm text-purple-400"> <span class="inline-block mr-1">🎺</span> <span itemprop="jobTitle">Trumpet</span> </p> <div class="hidden"> <p itemprop="description">Adam is a founding member of Sinister Dexter and is glad to be playing with them again after a brief 14-year hiatus. Starting as a trombone player in bands and orchestras in 9th grade, Adam picked up trumpet after college and decided to stick with it when SinDex began, both because the band already had two great trombone players and because it's way easier to carry a trumpet around. In the rare moments when he's not busy helping raise a kid or designing voice experiences for a large tech company, he likes to arrange and occasionally compose original music.</p> </div> <div class="rounded-lg p-3 -translate-x-1/2 absolute bg-gray-900 group-hover:opacity-100 left-1/2 mt-2 opacity-0 pointer-events-none shadow-xl transform transition-opacity w-64 z-20"> <p class="text-xs mb-2 text-gray-300">Adam is a founding member of Sinister Dexter and is glad to be playing with them again after a brief...</p> </div> </div> <div class="text-center group" itemscope itemtype="https://schema.org/Person" itemprop="member"> <meta content="https://sinisterdexter.net#lori-karns" itemprop="@id"> <meta content="Sinister Dexter" itemprop="affiliation"> <div class="relative inline-block mb-3"> <picture> <source srcset="images/members/lori.webp" type="image/webp"> <img alt="Lori Karns" class="transition-all rounded-full border-2 border-purple-500 group-hover:border-purple-400 group-hover:scale-105 h-32 mx-auto object-cover w-32" src="images/members/lori.jpg" itemprop="image"> </picture> </div> <h3 class="font-bold text-white" itemprop="name">Lori Karns</h3> <p class="text-sm text-purple-400"> <span class="inline-block mr-1">🎺</span> <span itemprop="jobTitle">Trumpet</span> </p> <div class="hidden"> <p itemprop="description"></p> </div> <div class="rounded-lg p-3 -translate-x-1/2 absolute bg-gray-900 group-hover:opacity-100 left-1/2 mt-2 opacity-0 pointer-events-none shadow-xl transform transition-opacity w-64 z-20"> <p class="text-xs mb-2 text-gray-300"></p> </div> </div> </div> </div> </section> <section class="px-6 py-20 bg-gradient-to-b from-black to-purple-950/30" id="notable-venues" aria-labelledby="venues-heading"> <div class="mx-auto max-w-6xl"> <h2 class="text-center bebas gradient-text md:text-6xl text-5xl mb-12" id="venues-heading">Notable Venues</h2> <div class="text-center mx-auto max-w-3xl gap-6 grid grid-cols-2 md:grid-cols-3"> <div class="group"> <div class="text-purple-400 group-hover:scale-110 mb-3 text-4xl transition-transform"> <i class="fas fa-music"></i> </div> <h3 class="font-bold text-white mb-1 text-lg">Club Fox</h3> <p class="text-sm text-gray-400">Redwood City</p> </div> <div class="group"> <div class="text-purple-400 group-hover:scale-110 mb-3 text-4xl transition-transform"> <i class="fas fa-beer"></i> </div> <h3 class="font-bold text-white mb-1 text-lg">Lost & Found Saloon</h3> <p class="text-sm text-gray-400">San Francisco</p> </div> <div class="group"> <div class="text-purple-400 group-hover:scale-110 mb-3 text-4xl transition-transform"> <i class="fas fa-microphone"></i> </div> <h3 class="font-bold text-white mb-1 text-lg">Blondie's Bar</h3> <p class="text-sm text-gray-400">San Francisco</p> </div> <div class="group"> <div class="text-purple-400 group-hover:scale-110 mb-3 text-4xl transition-transform"> <i class="fas fa-tree"></i> </div> <h3 class="font-bold text-white mb-1 text-lg">Stafford Park</h3> <p class="text-sm text-gray-400">Redwood City</p> </div> <div class="group"> <div class="text-purple-400 group-hover:scale-110 mb-3 text-4xl transition-transform"> <i class="fas fa-guitar"></i> </div> <h3 class="font-bold text-white mb-1 text-lg">Quarter Note</h3> <p class="text-sm text-gray-400">Sunnyvale</p> </div> <div class="group"> <div class="text-purple-400 group-hover:scale-110 mb-3 text-4xl transition-transform"> <i class="fas fa-coffee"></i> </div> <h3 class="font-bold text-white mb-1 text-lg">Blue Chalk Cafe</h3> <p class="text-sm text-gray-400">Palo Alto</p> </div> </div> <div class="text-center mt-12"> <p class="text-gray-400 italic text-lg"> Plus festivals, private events, and outdoor concert series throughout the Bay Area </p> </div> </div> </section> <section class="px-6 py-20 bg-gradient-to-b from-black to-purple-950/30" id="connect" aria-labelledby="connect-heading"> <div class="text-center mx-auto max-w-4xl"> <h2 class="bebas gradient-text md:text-6xl text-5xl mb-12" id="connect-heading">Stay Connected</h2> <div class="mb-12"> <h3 class="font-bold text-2xl mb-6">Follow Us</h3> <div class="flex justify-center gap-6"> <a class="rounded focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-purple-400 hover:text-purple-400 social-icon text-4xl" href="https://www.instagram.com/sindex_band" target="_blank" itemprop="sameAs" aria-label="Follow Sinister Dexter on Instagram"> <svg class="h-6 w-6" fill="currentColor" viewBox="0 0 24 24" aria-hidden="true" xmlns="http://www.w3.org/2000/svg" role="img"> <path d="M7.0301.084c-1.2768.0602-2.1487.264-2.911.5634-.7888.3075-1.4575.72-2.1228 1.3877-.6652.6677-1.075 1.3368-1.3802 2.127-.2954.7638-.4956 1.6365-.552 2.914-.0564 1.2775-.0689 1.6882-.0626 4.947.0062 3.2586.0206 3.6671.0825 4.9473.061 1.2765.264 2.1482.5635 2.9107.308.7889.72 1.4573 1.388 2.1228.6679.6655 1.3365 1.0743 2.1285 1.38.7632.295 1.6361.4961 2.9134.552 1.2773.056 1.6884.069 4.9462.0627 3.2578-.0062 3.668-.0207 4.9478-.0814 1.28-.0607 2.147-.2652 2.9098-.5633.7889-.3086 1.4578-.72 2.1228-1.3881.665-.6682 1.0745-1.3378 1.3795-2.1284.2957-.7632.4966-1.636.552-2.9124.056-1.2809.0692-1.6898.063-4.948-.0063-3.2583-.021-3.6668-.0817-4.9465-.0607-1.2797-.264-2.1487-.5633-2.9117-.3084-.7889-.72-1.4568-1.3876-2.1228C21.2982 1.33 20.628.9208 19.8378.6165 19.074.321 18.2017.1197 16.9244.0645 15.6471.0093 15.236-.005 11.977.0014 8.718.0076 8.31.0215 7.0301.0839m.1402 21.6932c-1.17-.0509-1.8053-.2453-2.2287-.408-.5606-.216-.96-.4771-1.3819-.895-.422-.4178-.6811-.8186-.9-1.378-.1644-.4234-.3624-1.058-.4171-2.228-.0595-1.2645-.072-1.6442-.079-4.848-.007-3.2037.0053-3.583.0607-4.848.05-1.169.2456-1.805.408-2.2282.216-.5613.4762-.96.895-1.3816.4188-.4217.8184-.6814 1.3783-.9003.423-.1651 1.0575-.3614 2.227-.4171 1.2655-.06 1.6447-.072 4.848-.079 3.2033-.007 3.5835.005 4.8495.0608 1.169.0508 1.8053.2445 2.228.408.5608.216.96.4754 1.3816.895.4217.4194.6816.8176.9005 1.3787.1653.4217.3617 1.056.4169 2.2263.0602 1.2655.0739 1.645.0796 4.848.0058 3.203-.0055 3.5834-.061 4.848-.051 1.17-.245 1.8055-.408 2.2294-.216.5604-.4763.96-.8954 1.3814-.419.4215-.8181.6811-1.3783.9-.4224.1649-1.0577.3617-2.2262.4174-1.2656.0595-1.6448.072-4.8493.079-3.2045.007-3.5825-.006-4.848-.0608M16.953 5.5864A1.44 1.44 0 1 0 18.39 4.144a1.44 1.44 0 0 0-1.437 1.4424M5.8385 12.012c.0067 3.4032 2.7706 6.1557 6.173 6.1493 3.4026-.0065 6.157-2.7701 6.1506-6.1733-.0065-3.4032-2.771-6.1565-6.174-6.1498-3.403.0067-6.156 2.771-6.1496 6.1738M8 12.0077a4 4 0 1 1 4.008 3.9921A3.9996 3.9996 0 0 1 8 12.0077"/> </svg> </a> <a class="rounded focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-purple-400 hover:text-purple-400 social-icon text-4xl" href="https://x.com/sindexband" target="_blank" itemprop="sameAs" aria-label="Follow Sinister Dexter on X (Twitter)"> <svg class="h-6 w-6" fill="currentColor" viewBox="0 0 24 24" aria-hidden="true" xmlns="http://www.w3.org/2000/svg" role="img"> <path d="M18.901 1.153h3.68l-8.04 9.19L24 22.846h-7.406l-5.8-7.584-6.638 7.584H.474l8.6-9.83L0 1.154h7.594l5.243 6.932ZM17.61 20.644h2.039L6.486 3.24H4.298Z"/> </svg> </a> <a class="rounded focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-purple-400 hover:text-purple-400 social-icon text-4xl" href="https://facebook.com/SinDex.Band/" target="_blank" itemprop="sameAs" aria-label="Follow Sinister Dexter on Facebook"> <svg class="h-6 w-6" fill="currentColor" viewBox="0 0 24 24" aria-hidden="true" xmlns="http://www.w3.org/2000/svg" role="img"> <path d="M9.101 23.691v-7.98H6.627v-3.667h2.474v-1.58c0-4.085 1.848-5.978 5.858-5.978.401 0 .955.042 1.468.103a8.68 8.68 0 0 1 1.141.195v3.325a8.623 8.623 0 0 0-.653-.036 26.805 26.805 0 0 0-.733-.009c-.707 0-1.259.096-1.675.309a1.686 1.686 0 0 0-.679.622c-.258.42-.374.995-.374 1.752v1.297h3.919l-.386 3.667h-3.533v7.98h2.959A15.999 15.999 0 0 0 24 12.073c0-8.838-7.162-16-16-16S-8 3.235-8 12.073c0 7.662 5.38 14.048 12.561 15.618"/> </svg> </a> <a class="rounded focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-purple-400 hover:text-purple-400 social-icon text-4xl" href="https://www.youtube.com/@SinDexBand" target="_blank" itemprop="sameAs" aria-label="Subscribe to Sinister Dexter on YouTube"> <svg class="h-6 w-6" fill="currentColor" viewBox="0 0 24 24" aria-hidden="true"> <path d="M23.498 6.186a3.016 3.016 0 0 0-2.122-2.136C19.505 3.545 12 3.545 12 3.545s-7.505 0-9.377.505A3.017 3.017 0 0 0 .502 6.186C0 8.07 0 12 0 12s0 3.93.502 5.814a3.016 3.016 0 0 0 2.122 2.136c1.871.505 9.376.505 9.376.505s7.505 0 9.377-.505a3.015 3.015 0 0 0 2.122-2.136C24 15.93 24 12 24 12s0-3.93-.502-5.814zM9.545 15.568V8.432L15.818 12l-6.273 3.568z"/> </svg> </a> </div> </div> <div class="mb-12"> <h3 class="font-bold text-2xl mb-6">Stream Our Music</h3> <div class="flex justify-center flex-col gap-4 sm:flex-row"> <a class="flex items-center justify-center rounded-full hover:scale-105 px-6 py-3 transition-all bg-green-600 hover:bg-green-700" href="https://open.spotify.com/artist/06ZpJFLW4Kz4EAlUorzrFu" target="_blank" aria-label="Listen to Sinister Dexter on Spotify"> <svg class="h-6 w-6 mr-2" fill="currentColor" viewBox="0 0 24 24" aria-hidden="true"> <path d="M12 0C5.4 0 0 5.4 0 12s5.4 12 12 12 12-5.4 12-12S18.66 0 12 0zm5.521 17.34c-.24.359-.66.48-1.021.24-2.82-1.74-6.36-2.101-10.561-1.141-.418.122-.779-.179-.899-.539-.12-.421.18-.78.54-.9 4.56-1.021 8.52-.6 11.64 1.32.42.18.479.659.301 1.02zm1.44-3.3c-.301.42-.841.6-1.262.3-3.239-1.98-8.159-2.58-11.939-1.38-.479.12-1.02-.12-1.14-.6-.12-.48.12-1.021.6-1.141C9.6 9.9 15 10.561 18.72 12.84c.361.181.54.78.241 1.2zm.12-3.36C15.24 8.4 8.82 8.16 5.16 9.301c-.6.179-1.2-.181-1.38-.721-.18-.601.18-1.2.72-1.381 4.26-1.26 11.28-1.02 15.721 1.621.539.3.719 1.02.419 1.56-.299.421-1.02.599-1.559.3z"/> </svg> <span class="font-semibold text-lg">Spotify</span> </a> <a class="flex items-center justify-center rounded-full hover:scale-105 px-6 py-3 transition-all bg-gray-700 hover:bg-gray-600" href="https://music.apple.com/us/artist/sinister-dexter/30663361" target="_blank" aria-label="Listen to Sinister Dexter on Apple Music"> <svg class="h-6 w-6 mr-2" fill="currentColor" viewBox="0 0 24 24" aria-hidden="true"> <path d="M12.152 6.896c-.948 0-2.415-1.078-3.96-1.04-2.04.027-3.91 1.183-4.961 3.014-2.117 3.675-.546 9.103 1.519 12.09 1.013 1.454 2.208 3.09 3.792 3.039 1.52-.065 2.09-.987 3.935-.987 1.831 0 2.35.987 3.96.948 1.637-.026 2.676-1.48 3.676-2.948 1.156-1.688 1.636-3.325 1.662-3.415-.039-.013-3.182-1.221-3.22-4.857-.026-3.04 2.48-4.494 2.597-4.559-1.429-2.09-3.623-2.324-4.39-2.376-2-.156-3.675 1.09-4.61 1.09zM15.53 3.83c.843-1.012 1.4-2.427 1.245-3.83-1.207.052-2.662.805-3.532 1.818-.78.896-1.454 2.338-1.273 3.714 1.338.104 2.715-.688 3.559-1.701"/> </svg> <span class="font-semibold text-lg">Apple Music</span> </a> </div> </div> <div class="rounded-lg bg-gray-900/50 border border-purple-500/30 p-8 mb-12"> <h3 class="font-bold text-2xl mb-4">Join Our Mailing List</h3> <p class="mb-6">Get updates on upcoming shows, new music, and exclusive content!</p> <div class="text-center mx-auto max-w-md"> <p class="text-gray-300 mb-4">Send an email to subscribe:</p> <a class="transition-all rounded-full hover:scale-105 py-3 items-center px-6 inline-flex font-semibold bg-purple-600 hover:bg-purple-700" href="mailto:gigs-subscribe@sinisterdexter.net?subject=Subscribe&body=Please%20add%20me%20to%20the%20Sinister%20Dexter%20mailing%20list!"> <svg class="h-5 w-5 mr-2" fill="currentColor" viewBox="0 0 20 20"> <path d="M2.003 5.884L10 9.882l7.997-3.998A2 2 0 0016 4H4a2 2 0 00-1.997 1.884z"/> <path d="M18 8.118l-8 4-8-4V14a2 2 0 002 2h12a2 2 0 002-2V8.118z"/> </svg> Subscribe to Mailing List </a> <p class="text-sm text-gray-400 mt-4"> Or email: <span class="font-mono">gigs-subscribe@sinisterdexter.net</span> </p> </div> </div> <div class="rounded-lg bg-purple-900/30 p-6"> <h3 class="font-bold text-2xl mb-4">Book the Band</h3> <p class="mb-4">Available for clubs, festivals, corporate events, and private parties</p> <a class="transition-all rounded-full hover:scale-105 py-3 items-center px-6 inline-flex bg-amber-500 font-bold hover:bg-amber-600 text-black" href="mailto:booking@sinisterdexter.net?subject=Booking%20Inquiry"> <svg class="h-5 w-5 mr-2" fill="currentColor" viewBox="0 0 20 20"> <path d="M2 3a1 1 0 011-1h2.153a1 1 0 01.986.836l.74 4.435a1 1 0 01-.54 1.06l-1.548.773a11.037 11.037 0 006.105 6.105l.774-1.548a1 1 0 011.059-.54l4.435.74a1 1 0 01.836.986V17a1 1 0 01-1 1h-2C7.82 18 2 12.18 2 5V3z"/> </svg> Contact for Booking </a> <p class="text-sm text-gray-400 mt-4"> Email: <span class="font-mono">booking@sinisterdexter.net</span> </p> </div> </div> </section> </main> <footer class="px-6 bg-black border-gray-800 border-t py-8"> <div class="text-center mx-auto max-w-7xl"> <p class="text-gray-400 mb-2">© 2025 Sinister Dexter. All rights reserved.</p> <p class="text-sm text-gray-500">The Right to Be Funky™</p> </div> </footer> <div class="left-0 bottom-0 fixed right-0 z-50" id="bottom-player" style="padding-bottom:env(safe-area-inset-bottom);background:linear-gradient(to bottom,rgba(17,24,39,.85) 0,rgba(17,24,39,.95) 30%,#111827 100%);backdrop-filter:blur(10px);-webkit-backdrop-filter:blur(10px);border-top:2px solid #9333ea;box-shadow:0 -10px 40px rgba(0,0,0,.5)"> <div class="mx-auto container md:px-4 md:py-4 px-3 py-3"> <div class="flex items-center gap-2 md:gap-3"> <button class="transition-all p-3 bg-gradient-to-r duration-300 focus:outline-none focus:ring-2 focus:ring-purple-400 from-purple-600 hover:from-purple-700 hover:to-purple-900 md:p-4 rounded-full to-purple-800" aria-label="Play or pause music" id="play-pause-btn"> <svg class="h-6 w-6 md:h-8 md:w-8 text-white" fill="currentColor" viewBox="0 0 20 20" id="play-icon"> <path d="M6.3 2.841A1.5 1.5 0 004 4.11v11.78a1.5 1.5 0 002.3 1.269l9.344-5.89a1.5 1.5 0 000-2.538L6.3 2.84z"/> </svg> <svg class="hidden text-white h-6 md:h-8 md:w-8 w-6" fill="currentColor" viewBox="0 0 20 20" id="pause-icon"> <path d="M5.75 3a.75.75 0 00-.75.75v12.5c0 .414.336.75.75.75h1.5a.75.75 0 00.75-.75V3.75A.75.75 0 007.25 3h-1.5zM12.75 3a.75.75 0 00-.75.75v12.5c0 .414.336.75.75.75h1.5a.75.75 0 00.75-.75V3.75a.75.75 0 00-.75-.75h-1.5z"/> </svg> </button> <button class="text-gray-400 focus:outline-none focus:ring-2 focus:ring-purple-400 hover:text-white p-2 rounded transition-colors" aria-label="Previous track" id="prev-btn"> <svg class="h-5 w-5" fill="currentColor" viewBox="0 0 20 20"> <path d="M8.445 14.832A1 1 0 0010 14v-2.798l5.445 3.63A1 1 0 0017 14V6a1 1 0 00-1.555-.832L10 8.798V6a1 1 0 00-1.555-.832l-6 4a1 1 0 000 1.664l6 4z"/> </svg> </button> <button class="text-gray-400 focus:outline-none focus:ring-2 focus:ring-purple-400 hover:text-white p-2 rounded transition-colors" aria-label="Next track" id="next-btn"> <svg class="h-5 w-5" fill="currentColor" viewBox="0 0 20 20"> <path d="M4.555 5.168A1 1 0 003 6v8a1 1 0 001.555.832L10 11.202V14a1 1 0 001.555.832l6-4a1 1 0 000-1.664l-6-4A1 1 0 0010 6v2.798l-5.445-3.63z"/> </svg> </button> <div class="text-center flex-1 md:px-3 min-w-0 px-2"> <div class="font-semibold md:text-lg text-base text-white truncate" id="track-title">No track selected</div> <div class="text-sm text-gray-400 hidden sm:block" id="track-subtitle">Click a track to play</div> </div> <div class="flex items-center gap-2 md:gap-3 flex-1"> <span class="text-xs hidden md:block text-gray-400 w-10 text-right" id="current-time">0:00</span> <div class="relative cursor-pointer bg-gray-700 flex-1 h-1.5 rounded-full" id="progress-container"> <div class="h-full absolute bg-gradient-to-r from-purple-600 left-0 rounded-full to-amber-500 top-0" id="progress-bar" style="width:0%"></div> </div> <span class="text-xs hidden md:block text-gray-400 w-10" id="duration">0:00</span> </div> </div> </div> </div> <audio id="bottom-audio-element" preload="metadata"></audio> <script>const tracks=[{num:1,title:"Broadband Connection",artist:"Sinister Dexter",duration:"3:01",filename:"Sinister_Dexter-01-Broadband_Connection.mp3",src:"https://cdn.sinister-dexter.com/music/Sinister_Dexter-01-Broadband_Connection.mp3"},{num:2,title:"Freeman",artist:"Sinister Dexter",duration:"3:03",filename:"Sinister_Dexter-02-Freeman.mp3",src:"https://cdn.sinister-dexter.com/music/Sinister_Dexter-02-Freeman.mp3"},{num:3,title:"Chasing the Clouds Away",artist:"Sinister Dexter",duration:"3:12",filename:"Sinister_Dexter-02-Chasing_the_Clouds_Away.mp3",src:"https://cdn.sinister-dexter.com/music/Sinister_Dexter-02-Chasing_the_Clouds_Away.mp3"},{num:4,title:"Morenita",artist:"Sinister Dexter",duration:"3:26",filename:"Sinister_Dexter-03-Morenita.mp3",src:"https://cdn.sinister-dexter.com/music/Sinister_Dexter-03-Morenita.mp3"},{num:5,title:"One-Eyed Jack",artist:"Sinister Dexter",duration:"4:15",filename:"Sinister_Dexter-03-One-Eyed_Jack.mp3",src:"https://cdn.sinister-dexter.com/music/Sinister_Dexter-03-One-Eyed_Jack.mp3"},{num:6,title:"Are You Ready to Dance",artist:"Sinister Dexter",duration:"3:26",filename:"Sinister_Dexter-04-Are_You_Ready_to_Dance.mp3",src:"https://cdn.sinister-dexter.com/music/Sinister_Dexter-04-Are_You_Ready_to_Dance.mp3"},{num:7,title:"Sinister Dexter",artist:"Sinister Dexter",duration:"3:26",filename:"Sinister_Dexter-04-Sinister_Dexter.mp3",src:"https://cdn.sinister-dexter.com/music/Sinister_Dexter-04-Sinister_Dexter.mp3"},{num:8,title:"Two Before Nine",artist:"Sinister Dexter",duration:"4:05",filename:"Sinister_Dexter-05-Two_Before_Nine.mp3",src:"https://cdn.sinister-dexter.com/music/Sinister_Dexter-05-Two_Before_Nine.mp3"},{num:9,title:"Can't Take My Eyes Off You",artist:"Sinister Dexter",duration:"4:42",filename:"Sinister_Dexter-06-Cant_Take_My_Eyes_Off_You.mp3",src:"https://cdn.sinister-dexter.com/music/Sinister_Dexter-06-Cant_Take_My_Eyes_Off_You.mp3"},{num:10,title:"Jack in the Box",artist:"Sinister Dexter",duration:"2:35",filename:"Sinister_Dexter-09-Jack_in_the_Box.mp3",src:"https://cdn.sinister-dexter.com/music/Sinister_Dexter-09-Jack_in_the_Box.mp3"},{num:11,title:"Him or Me",artist:"Sinister Dexter",duration:"4:49",filename:"Sinister_Dexter-10-Him_or_Me.mp3",src:"https://cdn.sinister-dexter.com/music/Sinister_Dexter-10-Him_or_Me.mp3"},{num:12,title:"Swim on Over",artist:"Sinister Dexter",duration:"4:08",filename:"Sinister_Dexter-11-Swim_on_Over.mp3",src:"https://cdn.sinister-dexter.com/music/Sinister_Dexter-11-Swim_on_Over.mp3"},{num:13,title:"Better Left Unsaid",artist:"Sinister Dexter",duration:"4:25",filename:"Sinister_Dexter-12-Better_Left_Unsaid.mp3",src:"https://cdn.sinister-dexter.com/music/Sinister_Dexter-12-Better_Left_Unsaid.mp3"},{num:14,title:"Another Drink",artist:"Sinister Dexter",duration:"3:04",filename:"Sinister_Dexter-13-Another_Drink.mp3",src:"https://cdn.sinister-dexter.com/music/Sinister_Dexter-13-Another_Drink.mp3"},{num:15,title:"Show Us How to Get Down (Live)",artist:"Sinister Dexter",duration:"5:53",filename:"Sinister_Dexter-01-Show_Us_How_to_Get_Down.mp3",src:"https://cdn.sinister-dexter.com/music/Sinister_Dexter-01-Show_Us_How_to_Get_Down.mp3"},{num:16,title:"Show Us How to Get Down (Studio)",artist:"Sinister Dexter",duration:"5:58",filename:"Sinister_Dexter-08-Show_Us_How_to_Get_Down.mp3",src:"https://cdn.sinister-dexter.com/music/Sinister_Dexter-08-Show_Us_How_to_Get_Down.mp3"},{num:17,title:"Show Us How to Get Down (Alt)",artist:"Sinister Dexter",duration:"6:30",filename:"Sinister_Dexter-Show_Us_How_to_Get_Down.mp3",src:"https://cdn.sinister-dexter.com/music/Sinister_Dexter-Show_Us_How_to_Get_Down.mp3"},{num:18,title:"Son of a Preacher Man",artist:"Sinister Dexter",duration:"3:03",filename:"Sinister_Dexter-Son_of_a_Preacher_Man.mp3",src:"https://cdn.sinister-dexter.com/music/Sinister_Dexter-Son_of_a_Preacher_Man.mp3"},{num:19,title:"Dream Come True",artist:"Sinister Dexter",duration:"5:21",filename:"Sinister_Dexter-Dream_Come_True.mp3",src:"https://cdn.sinister-dexter.com/music/Sinister_Dexter-Dream_Come_True.mp3"},{num:20,title:"What is Hip",artist:"Sinister Dexter",duration:"5:51",filename:"Sinister_Dexter-What_is_Hip.mp3",src:"https://cdn.sinister-dexter.com/music/Sinister_Dexter-What_is_Hip.mp3"},{num:21,title:"The Letter",artist:"Sinister Dexter",duration:"3:58",filename:"Sinister_Dexter-The_Letter.mp3",src:"https://cdn.sinister-dexter.com/music/Sinister_Dexter-The_Letter.mp3"}];!function(){let e=-1,t=!1;const i=document.getElementById("bottom-audio-element"),n=document.getElementById("play-pause-btn"),r=document.getElementById("play-icon"),s=document.getElementById("pause-icon"),a=document.getElementById("prev-btn"),o=document.getElementById("next-btn"),m=document.getElementById("track-title"),c=document.getElementById("track-subtitle"),d=document.getElementById("current-time"),_=document.getElementById("duration"),u=document.getElementById("progress-bar"),l=document.getElementById("progress-container");function x(e){if(isNaN(e))return"0:00";return`${Math.floor(e/60)}:${Math.floor(e%60).toString().padStart(2,"0")}`}function S(t){if(t<0||t>=tracks.length)return;e=t;const n=tracks[t];i.src=n.src,m.textContent=n.title,c.textContent=n.artist,document.querySelectorAll(".track-item").forEach((e,i)=>{i===t?e.classList.add("active"):e.classList.remove("active")}),"mediaSession"in navigator&&(navigator.mediaSession.metadata=new MediaMetadata({title:n.title,artist:n.artist,album:"Sinister Dexter"}))}function D(){-1===e&&S(0),t?(i.pause(),t=!1,r.classList.remove("hidden"),s.classList.add("hidden")):(i.play(),t=!0,r.classList.add("hidden"),s.classList.remove("hidden"))}function p(){e>0&&(S(e-1),t&&i.play())}function h(){e<tracks.length-1&&(S(e+1),t&&i.play())}n?.addEventListener("click",D),a?.addEventListener("click",p),o?.addEventListener("click",h),document.querySelectorAll(".track-item").forEach((e,n)=>{e.addEventListener("click",function(){S(n),i.play(),t=!0,r.classList.add("hidden"),s.classList.remove("hidden")})}),i?.addEventListener("timeupdate",()=>{const e=i.currentTime/i.duration*100;u.style.width=`${e}%`,d.textContent=x(i.currentTime)}),i?.addEventListener("loadedmetadata",()=>{_.textContent=x(i.duration)}),i?.addEventListener("ended",()=>{h()}),l?.addEventListener("click",e=>{const t=l.getBoundingClientRect(),n=(e.clientX-t.left)/t.width;i.currentTime=n*i.duration}),"mediaSession"in navigator&&(navigator.mediaSession.setActionHandler("play",D),navigator.mediaSession.setActionHandler("pause",D),navigator.mediaSession.setActionHandler("previoustrack",p),navigator.mediaSession.setActionHandler("nexttrack",h)),tracks.length>0&&S(0)}()</script></body> </html>
