+++ template/partials/head/styles.hbs
@@ -138,12 +138,9 @@
 
 <!-- ============================================= -->
 <!-- STYLESHEETS                                  -->
-<!-- External CSS dependencies                    -->
+<!-- Production Tailwind CSS                      -->
 <!-- ============================================= -->
-<!-- ⚠️ WARNING: Tailwind CDN is for DEVELOPMENT ONLY! -->
-<!-- For production, install Tailwind CSS via PostCSS or Tailwind CLI: -->
-<!-- https://tailwindcss.com/docs/installation -->
-<script src="https://cdn.tailwindcss.com"></script>
+<link rel="stylesheet" href="styles.css">
 
 <!-- ============================================= -->
 <!-- FONT LOADING                                 -->
